% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{add_missing_fn012}
\alias{add_missing_fn012}
\title{Add FN012 record for missing SPC-GRPs}
\usage{
add_missing_fn012(fn012, fn123)
}
\arguments{
\item{fn012}{Dataframe containing FN012 Sampling specs and keys
"PRJ_CD", "SPC", "GRP"}

\item{fn123}{Dataframe containing catch count data and keys
"PRJ_CD", "SPC", "GRP"}
}
\value{
dataframe
}
\description{
This function adds records to the FN012 table for PRJ_CD-SPC-GRP
combinations that exist in the FN123 table (they were caught), but
are not currently in the FN012 (Sampling Spec) table.  This is a
helper function used by \code{\link{populate_fn012}} and is not
intended by called directly by users.
}
\seealso{
populate_fn012
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
