% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_year_column.R
\name{add_year_column}
\alias{add_year_column}
\title{Add year to fishnet-like dataframe}
\usage{
add_year_column(fn_data, colname = "YEAR", silent = FALSE)
}
\arguments{
\item{fn_data}{\itemize{
\item data frame with the column 'prj_cd'
}}

\item{silent}{\itemize{
\item should the warnings be suppressed
}}
}
\value{
data frame
}
\description{
The add_year_col is used by many of the functions glfishr but can
also be used independently to add a year column to a "FishNet-2"
like dataframe that contains a PRJ_CD column.  If the data frame
does not contain column "PRJ_CD" or "prj_cd" an error will be
thrown. If the data frame already contains a Year column, it will
be returned unchanged.
}
\details{
Add year to fishnet-like dataframe
}
\author{
R. Adam Cottrill
}
