% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{append_data}
\alias{append_data}
\title{Append data-frame to Access table}
\usage{
append_data(
  dbase,
  trg_table,
  data,
  verbose = T,
  verbose_sqlsave = F,
  append = T,
  safer = T
)
}
\arguments{
\item{dbase}{\itemize{
\item path to our target database
}}

\item{trg_table}{\itemize{
\item the table in the target data to insert data into
}}

\item{data}{\itemize{
\item data-frame to append in the target table
}}

\item{verbose}{\itemize{
\item should the number of records and table be printed?
}}

\item{verbose_sqlsave}{\itemize{
\item Boolean. Default=FALSE. This value is
passed through to RODBC::sqlsave.  Set to True to see all of
the sql statements passed to target database.  Useful for
debugging.
}}

\item{append}{\itemize{
\item append to and existing table, or drop and create it?
}}

\item{safer}{\itemize{
\item passed to RODBC::sqlSave
}}
}
\value{
status of closed RODBC connection.
}
\description{
This function will attempt to append data contained in a data-frame
to a table in the target database.  If verbose=TRUE, a statement
including the number of records effected will be reported to the
console.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
