% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{assign_fn012_sizesam}
\alias{assign_fn012_sizesam}
\title{Populate FN012.SIZSAM based}
\usage{
assign_fn012_sizesam(fn012, fn124, fn125)
}
\arguments{
\item{fn012}{\itemize{
\item dataframe containing FN012 sampling spec. data
}}

\item{fn124}{\itemize{
\item dataframe containing FN124 length tally data
}}

\item{fn125}{\itemize{
\item dataframe containing FN125 biological data
}}
}
\value{
dataframe
}
\description{
This function updates the BIOSAM values in the FN012 table based
on the presence of SPC-GRP records in the associated FN124 and
FN125.
}
\seealso{
fill_missing_fn012_limits, populate_fn012
}
\author{
Arthur Bonsall \email{arthur.bonsall@ontario.ca}
}
