% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{augment_fn012}
\alias{augment_fn012}
\title{Augment the FN012 data}
\usage{
augment_fn012(fn011, fn012, fn123, prune_fn012, source)
}
\arguments{
\item{fn011}{\itemize{
\item dataframe contains fn011 data
}}

\item{fn012}{\itemize{
\item dataframe contains fn012 data fetched from the api
for given filters
}}

\item{fn123}{\itemize{
\item dataframe contains fn123 data fetched from the api
for given filters
}}

\item{prune_fn012}{\itemize{
\item boolean - should unused FN012 records be
removed from the table?
}}

\item{source}{\itemize{
\item string. Either "assessment" or "creel"
}}
}
\value{
dataframe
}
\description{
the fn012 data downloaded from the api, may be incomplete, missing
species and groups that that were actually encountered in the
associated  projects. This function adds records to the FN012 table
in these instances. If the missing spc-grp combinations can be
found in the FN012 protocols table, those values are used,
otherwise default values for the species are used instead.    If
prune=TRUE SPC-GRP combinations that appear in the FN012 table,
but not in the FN123 table, are dropped from the FN012 before it
is returned.
}
\seealso{
populate_fn012
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
