% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceed_limit.R
\name{exceed_limit}
\alias{exceed_limit}
\title{Exceed maximum number of API calls}
\usage{
exceed_limit(get_fun, prj_cds, extra_filters_as_list = NULL)
}
\arguments{
\item{get_fun}{Any \code{glfishr} function that begins with \code{get_}}

\item{prj_cds}{A list of PRJ_CD strings or a column containing each PRJ_CD of interest}

\item{extra_filters_as_list}{An optional list of additional filters to pass along to \code{get_fun}}
}
\value{
dataframe
}
\description{
The series of \code{glfishr} \verb{get_**} functions each retrieve a maximum number of results
when the function is called. Use \code{exceed_limit} to retrieve a dataframe that is longer
than this limit. \code{exceed_limit} applies your chosen get function to each PRJ_CD of
interest and binds those results into a new dataframe.
}
\examples{
FN011 <- get_FN011(list(protocol = "BSM", lake = "HU",
prj_cd__not = "LHR_IA17_819", prj_cd__not_like = "BM"))

my_projects <- list("LHA_IA19_802", "LHA_IA19_810", "LHA_IA19_811",
"LHA_IA19_812", "LHA_IA19_813", "LHA_IA19_814")

FN122 <- exceed_limit(get_FN122, FN011$PRJ_CD)

FN122_2019 <- exceed_limit(get_FN122, my_projects)

FN122_NA1 <- exceed_limit(get_FN122, FN011$PRJ_CD, list(gr = "NA1"))

FN125_NA1 <- exceed_limit(get_FN125, FN011$PRJ_CD,
list(spc = c(121, 331, 334)))
}
