% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{fetch_fn012_protocol_data}
\alias{fetch_fn012_protocol_data}
\title{Fetch FN012 data for lake-protocols}
\usage{
fetch_fn012_protocol_data(protocols)
}
\arguments{
\item{protocols}{\itemize{
\item dataframe with the columns "PRJ_CD", "LAKE",
"PROTOCOL"
}}
}
\value{
datarame
}
\description{
Given a dataframe with columns PRJ_CD, LAKE and PROTOCOL, fetch
the default fish sampling specs (FN012 records) for that protocol
in that lake. This is a helper function that is called by
get_or_augment_fn012.
}
\seealso{
get_or_augment_fn012
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
