% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sc012_protocol_data.R
\name{fetch_sc012_protocol_data}
\alias{fetch_sc012_protocol_data}
\title{Fetch SC012 data for a lake}
\usage{
fetch_sc012_protocol_data(protocols)
}
\arguments{
\item{protocols}{\itemize{
\item dataframe with the columns "PRJ_CD", "LAKE",
}}
}
\value{
datarame
}
\description{
Given a dataframe with columns PRJ_CD, and LAKE , fetch
the default fish sampling specs (FN012 records) for that creels
run in that lake. This is a helper function that is called by
get_or_augment_fn012 for creel projects.
}
\seealso{
get_or_augment_fn012
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
