% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tfat_lookups.R
\name{fetch_tfat_lookups}
\alias{fetch_tfat_lookups}
\title{Fetch the values from the tfat lookup api endpoint}
\usage{
fetch_tfat_lookups()
}
\value{
list
}
\description{
This is a helper function that is used to fetch the lookup values from tfat.
Tfat exposes a lookup endpoint that returns a json response that contains
key-value pairs for a number of elements associated with fish tagging - tag
types, tag colours, tagging agency etc.  This function calls that endpoint
and returns the results as a named list - one list for each key-value pair.
This function is not intended to be called directly, but is called by other,
higher-level convenience functions (~get_tag_colours()~).
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
