% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{fetching_report}
\alias{fetching_report}
\title{Report fetching activity to the console}
\usage{
fetching_report(table_name, verbose)
}
\arguments{
\item{table_name}{\itemize{
\item string. The name of the table to report on
}}

\item{verbose}{\itemize{
\item boolean. should the message be printed?
}}
}
\description{
A little helper function used by populate template to provide some
feedback to the user and report progress.  This function is not
intended to be called anywhere  other than from within
populate_template functions.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
