% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_fn012_limits.R
\name{fill_missing_fn012_limits}
\alias{fill_missing_fn012_limits}
\title{Fill missing FN012 size limits}
\usage{
fill_missing_fn012_limits(fn012)
}
\arguments{
\item{fn012}{\itemize{
\item dataframe containing FN012 sampling spec. data
}}
}
\value{
dataframe
}
\description{
The FN012 table has several fields that are used to bound
estimates of fish size and flag potential errors.  When new
records are added to the FN012 table, these values are null. This
function connects to the glis api endpoint and fetches the
attributes for the missing species and updates the corresponding
fields in the newly created fn012 records.
}
\seealso{
populate_fn012, assign_fn012_sizsam
}
\author{
Arthur Bonsall \email{arthur.bonsall@ontario.ca}
}
