% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ageprep1.R
\name{get_ageprep1}
\alias{get_ageprep1}
\title{Get age preparation 1 options - A list of age preparation 1 codes used in GLIS}
\usage{
get_ageprep1(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for age preparation 1 choices and returns
their labels, descriptions and whether they're in use. It fetches the entire
table of valid age preparation 1 codes - no other
filter parameters are currently available for this endpoint.
}
\details{
Note that age preparation 1 codes are valid for the second
character of the AGEMT field.

See
http://10.167.37.157/common/agepre1
for the full list of age preparation 1 code options
}
\examples{

ageprep1 <- get_ageprep1()
ageprep1_slugs <- get_ageprep1(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
