% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ageprep2.R
\name{get_ageprep2}
\alias{get_ageprep2}
\title{Get age preparation 2 options - A list of age preparation 2 codes used in GLIS}
\usage{
get_ageprep2(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for age preparation 2 choices and returns
their labels, descriptions and whether they're in use. It fetches the entire
table of valid age preparation 2 codes - no other
filter parameters are currently available for this endpoint.
}
\details{
Note that age preparation 2 codes are valid for the second
character of the AGEMT field.

See
https://intra.glis.mnr.gov.on.ca/common/ageprep2/
for the full list of age preparation 2 code options
}
\examples{

ageprep2 <- get_ageprep2()
ageprep2_slugs <- get_ageprep2(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
