% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agest.R
\name{get_agest}
\alias{get_agest}
\title{Get aging structure options - A list of aging structure codes used in GLIS}
\usage{
get_agest(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for aging structure choices and returns
their labels, descriptions and whether they're in use. It can fetch the entire
table of valid aging structures, or it accepts the filter parameter all=TRUE
to return all aging structures, including those no longer in use. No other
filter parameters are currently available for this endpoint.
}
\details{
Note that aging structures are valid for the AGEST field and the first
character of the AGEMT field.

See
http://10.167.37.157/common/aging_structures
for the full list of aging structure code options
}
\examples{

agest <- get_agest()
agest_all <- get_agest(list(all = TRUE))
agest_slugs <- get_agest(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
