% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_filters.R
\name{get_api_filters}
\alias{get_api_filters}
\title{Create global list of available API filters}
\usage{
get_api_filters(api_app = "fn_portal")
}
\arguments{
\item{api_app}{\itemize{
\item the name of the api application to fetch the
filters from. Defaults to 'fn_portal'
}}
}
\value{
list
}
\description{
This function connects to the openapi/swagger endpoint provided by
fn_portal and sc_portal api's and fetches all of the available
filters for each endpoint.  The filters are available in a global
list 'api_filters' which subsequently used by other functions -
check_filters, show_filters.  Generally, this function is not
intended to be called directly by the user.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
