% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{get_assessment_data}
\alias{get_assessment_data}
\title{connect the api and get data from assessment portal}
\usage{
get_assessment_data(filters, prune_fn012, verbose)
}
\arguments{
\item{filters}{\itemize{
\item list of filters passed to glis api endpoints
}}

\item{prune_fn012}{\itemize{
\item boolean. Should records in the FN012 table
without records in the FN123 be dropped?
}}

\item{verbose}{\itemize{
\item boolean.  Should progress be reproted in the console?
}}
}
\value{
named list
}
\description{
This is one of the workhorse functions used by populate_template.
This function connects to the assessment portal and fetches all of
the data selected by the filters argument.  Returns a named list-
names of the list correspond directly to names of the tables in
template database.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
