% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bottom_types.R
\name{get_bottom_types}
\alias{get_bottom_types}
\title{Get bottom types - A list of bottom types used in GLIS}
\usage{
get_bottom_types(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for lake bottom types and returns
their names, descriptions and whether they're in use. It can fetch
the entire table of accepted bottom type codes, or it accepts filter parameter
all=TRUE to return depreciated bottom type codes too. No other filter
parameters are currently available for this endpoint.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/bottom_types/
for the full list of bottom types
}
\examples{

bottom_types <- get_bottom_types()
all_bottom_types <- get_bottom_types(list(all = TRUE))
bottom_type_slugs <- get_bottom_types(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
