% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clip_codes.R
\name{get_clip_codes}
\alias{get_clip_codes}
\title{Get clip code options - A list of clip codes used in GLIS}
\usage{
get_clip_codes(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for clip code choices and returns
their labels, descriptions and whether they're in use. It fetches
the entire table of clip codes - no other filter
parameters are currently available for this endpoint. The same list of
clip codes are available for CLIPA and CLIPC.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/finclips/
for the full list of clip code options
}
\examples{

clip_codes <- get_clip_codes()
clip_code_slugs <- get_clip_codes(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
