% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cover_types.R
\name{get_cover_types}
\alias{get_cover_types}
\title{Get cover types - A list of cover types used in GLIS}
\usage{
get_cover_types(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for cover types and returns
their names, descriptions and whether they're in use. It can fetch
the entire table of accepted cover type codes, or it accepts filter parameter
all=TRUE to return depreciated cover type codes too. No other filter
parameters are currently available for this endpoint.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/cover_types/
for the full list of cover types
}
\examples{

cover_types <- get_cover_types()
all_cover_types <- get_cover_types(list(all = TRUE))
cover_type_slugs <- get_cover_types(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
