% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Get GLIS Credentials}
\usage{
get_credentials()
}
\description{
A helper function to prompt a user for their GLIS username and password.
Using rstudioapi creates a pop-up to enter credentials so that usernames
and passwords are not saved, and passwords are not visible.
}
