% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fisheries_offices.R
\name{get_fisheries_offices}
\alias{get_fisheries_offices}
\title{Get fisheries office options - A list of fisheries office codes used in GLIS}
\usage{
get_fisheries_offices(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for fisheries office choices and returns
their labels, descriptions and whether they're in use. It fetches
the entire table of accepted fisheries offices - no other filter
parameters are currently available for this endpoint. Fisheries office codes
represent the first three digits of project code (PRJ_CD).
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/fisheries_offices/
for the full list of fisheries_office options
}
\examples{

fisheries_offices <- get_fisheries_offices()
fisheries_office_slugs <- get_fisheries_offices(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
