% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN012.R
\name{get_FN012}
\alias{get_FN012}
\title{Get FN012 - Sampling specs (species and group) from FN_Portal API}
\usage{
get_FN012(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for FN012 records.
FN012 records contain a list of anticipated species for a project and the
likely length and weight ranges for those species. These constraints are used
by ProcVal to check that biodata for fish caught during the project is
reasonable. The FN012 records contain information like minimum and maximum
TLEN, FLEN, RWT, and K values for each species. This function takes an
optional filter list which can be used to return records based on several
attributes of the project such as protocol, lake, project code, and species.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn_012_list
for the full list of available filter keys (query parameters)
}
\examples{

fn012 <- get_FN012(list(lake = "HU"))

filters <- list(lake = "HU", prj_cd = c("LHA_IA21_823", "LHA_IA20_812"))
fn012 <- get_FN012(filters)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
