% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN013.R
\name{get_FN013}
\alias{get_FN013}
\title{Get FN013 - Gear Description Data FN_Portal API}
\usage{
get_FN013(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN013
records. FN013 records contain information about the defined gear
associated with a project.
}
\details{
Use \code{show_filters("fn013")} to see the full list of available filter
keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/}
and filter by "fn013" for additional information.
}
\examples{

fn013 <- get_FN013(list(lake = "ON", year = 2012))

fn013 <- get_FN013(list(prj_cd = "LOA_IA21_054"))
fn013 <- get_FN013(list(prj_cd = "L0A_IA21_054"), show_id = TRUE)
}
\author{
Jeremy Holden \email{jeremy.holden@ontario.ca}
}
