% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN026.R
\name{get_FN026}
\alias{get_FN026}
\title{Get FN026 - Spatial Strata Data FN_Portal API}
\usage{
get_FN026(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN026
records. FN026 records contain information about the spatial strata
associated with a project.
}
\details{
See
http://10.167.37.157/fn_portal/api/v1/redoc/#operation/fn_026_list
for the full list of available filter keys (query parameters)
}
\examples{

fn026 <- get_FN026(list(lake = "ON", year = 2012))

fn026 <- get_FN026(list(prj_cd = "LHA_IA19_812"))
fn026 <- get_FN026(list(prj_cd = "LHA_IA19_812"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
