% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN028.R
\name{get_FN028}
\alias{get_FN028}
\title{Get FN028 - Fishing Modes from FN_Portal API}
\usage{
get_FN028(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN028 records. FN028
records contain information about the fishing modes defined for a
project.  Fishing mode describes the gear, its orientation, and set
type.
}
\details{
Use \code{show_filters("fn028")} to see the full list of available filter
keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/}
and filter by "fn028" for additional information.
}
\examples{

fn028 <- get_FN028(list(lake = "ON", year = 2012))

fn028 <- get_FN028(list(prj_cd = "LHA_IA19_812"))
fn028 <- get_FN028(list(prj_cd = "LHA_IA19_812"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
