% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121_Electrofishing.R
\name{get_FN121_Electrofishing}
\alias{get_FN121_Electrofishing}
\title{Get FN121_Electrofishing - Electrofishing data from FN_Portal API}
\usage{
get_FN121_Electrofishing(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for FN121_Electrofishing
records. FN121_Electrofishing records contain information about
electrofishing sampling events, including shocking seconds, volts,
amps, and power used for each SAM. Other relevant details for the SAM
are found in the FN121 and optionally the FN121_GPS_Tracks tables.
This function takes an optional filter list which can be used to
return records based on attributes of the SAM including start and end
date and time, effort duration, gear, site depth and location as well as
attributes of the projects they are associated with such project
code, or part of the project code, lake, first year, last year,
protocol, etc.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn121electrofishing_list
for the full list of available filter keys (query parameters)
}
\examples{

TODO: Update with relevant examples when data exists in the portal

fn121_efish <- get_FN121_Electrofishing(list(lake = "ON", year = 2012))
fn121_efish <- get_FN121_Electrofishing(list(
  lake = "ER", protocol = "TWL",
  sidep__lte = 20, year__gte = 2010
))
filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN")
)
fn121_efish <- get_FN121_Electrofishing(filters)
fn121_efish <- get_FN121_Electrofishing(list(lake = "HU", prj_cd__like = "_003"))
fn121_efish <- get_FN121_Electrofishing(list(lake = "HU", prj_cd__like = "_003"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
