% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121_GPS.R
\name{get_FN121_GPS}
\alias{get_FN121_GPS}
\title{Get FN121_GPS - GPS data from FN_Portal API}
\usage{
get_FN121_GPS(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for FN121_GPS_Tracks
records. FN121_GPS_Tracks records contain GPS tracks for projects
where GPS data is recorded (e.g. trawls, electrofishing, etc.), including
the track ID, coordinates in decimal decrees, the timestamp and site depth.
Other relevant details for each SAM are found in the FN121 table.
This function takes an optional filter list which can be used to
return records based on attributes of the SAM including site depth, timestamp,
start and end date and time, effort duration, gear, site depth and location
as well as attributes of the projects they are associated with such project
code, or part of the project code, lake, first year, last year,
protocol, etc.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn121GpsTracks_list
for the full list of available filter keys (query parameters)
}
\examples{

TODO: Update with relevant examples when data exists in the portal

fn121_gps <- get_FN121_GPS(list(lake = "ON", year = 2012))
fn121_gps <- get_FN121_GPS(list(
  lake = "ER", protocol = "TWL",
  sidep__lte = 20, year__gte = 2010
))
filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN")
)
fn121_gps <- get_FN121_GPS(filters)
fn121_gps <- get_FN121_GPS(list(lake = "HU", prj_cd__like = "_003"))
fn121_gps <- get_FN121_GPS(list(lake = "HU", prj_cd__like = "_003"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
