% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121_Trawl.R
\name{get_FN121_Trawl}
\alias{get_FN121_Trawl}
\title{Get FN121_Trawl - Trawl data from FN_Portal API}
\usage{
get_FN121_Trawl(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for fn121trawl
records. fn121trawl records contain details on the vessel, its speed and direction,
and warp. Other relevant details for each SAM are found in the FN121 table.
This function takes an optional filter list which can be used to return records
based on attributes of the SAM including site depth, start and end date
and time, effort duration, gear, and location as well as attributes of the projects
they are associated with such as project code, or part of the project code, lake,
first year, last year, protocol, etc. This function can also take filters related to
the vessel/warp.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn121trawl_list
for the full list of available filter keys (query parameters)
}
\examples{

TODO: Update with relevant examples when more data exists in the portal

fn121_trawl <- get_FN121_Trawl(list(lake = "ER", year = 2018))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
