% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121Limno.R
\name{get_FN121limno}
\alias{get_FN121limno}
\title{Get FN121limno - Net set limnological data from FN_Portal API}
\usage{
get_FN121limno(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?  Defaults to TRUE to match the FISHNET-II  convention
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN121limno
records. FN121limno records contain limnological data associated
with individual net sets.  There is a 1:1 relationship between a
net set and a FN121Lino record.  The FN121limno record contains
information like disolved oxygen at the surface or at the gear at
the start and/or end of a sampling event.  This function takes an
optional filter list which can be used to return records based on
several attributes of the limnological data as well as atttributes
of the net set including set and lift date and
time, effort duration, gear, site depth and location as well as
attributes of the projects they are associated with such project
code, or part of the project code, lake, first year, last year,
protocol, etc.  The limnogical data can be joined back to the
associated net set in R using the function \code{\link{join_fn_fields}}
}
\details{
See See \href{http://10.167.37.157/fn_portal/redoc/#operation/fn121limno}{fn121limno}
or type \code{\link{show_filters}} to see
the full list of available filter keys (query parameters).
}
\examples{

# see the available filters that contain the string 'o2' (oxygen)
show_filters("fn121limno", "o2")

filter_list <- list(prj_cd = "LHA_IA14_801")
fn121limno <- get_FN121limno(filter_list)

# pass the same fitler to our FN121 to get the associated net sets:

fn121 <- get_FN121(filter_list)
joined <- join_fn_fields(fn121, fn121limno)
# net sets are now associated with their limnlogical data:
head(joined)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
