% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN123.R
\name{get_FN123}
\alias{get_FN123}
\title{Get FN123 - Catch Counts from FN_Portal API}
\usage{
get_FN123(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN123 records from
the FN_Portal. FN123 records contain information about catch
counts by species for each effort in a sample.  For most gill
netting projects this corresponds to catches within a single panel
of a particular mesh size within a net set (gang). Group (GRP) is
occasionally included to further sub-divide the catch into user
defined groups that are usually specific to the project.  This
function takes an optional filter list which can be used to return
record based on several attributes of the catch including species,
or group code but also attributes of the effort, the sample or the
project(s) that the catches were made in.
}
\details{
Use ~show_filters("fn123")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/
and filter by "fn123" for additional information.
}
\examples{

fn123 <- get_FN123(list(lake = "ON", year = 2012, spc = "334", gr = "GL"))

filters <- list(
  lake = "ER",
  protocol = "TWL",
  year = 2010,
  spc = c("331", "334"),
  sidep0__lte = 20
)
fn123 <- get_FN123(filters)

filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"),
  eff = "051",
  spc = "091"
)
fn123 <- get_FN123(filters)

filters <- list(lake = "HU", spc = "076", grp = "55")
fn123 <- get_FN123(filters)

fn123 <- get_FN123(list(prj_cd = "LHA_IA19_812"))
fn123 <- get_FN123(list(prj_cd = "LHA_IA19_812"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
