% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN123_NonFish.R
\name{get_FN123_NonFish}
\alias{get_FN123_NonFish}
\title{Get FN123_NonFish - Catch Counts of Non-Fish species from FN_Portal API}
\usage{
get_FN123_NonFish(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN123_NonFish records from
the FN_Portal. FN123_NonFish records contain information about catch
counts by non-fish species for each effort in a sample. This
function takes an optional filter list which can be used to return
record based on several attributes of the catch including species,
but also attributes of the effort, the sample or the
project(s) that the catches were made in.
}
\details{
See
http://10.167.37.157/fn_portal/api/v1/redoc/#operation/fn123nonfish_list
for the full list of available filter keys (query parameters)
}
\examples{

fn123_nonfish <- get_FN123_NonFish(list(lake = "ER", year = 2019))

fn123_nonfish <- get_FN123_NonFish(list(prj_cd = "LEA_IA19_SHA"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
