% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN124.R
\name{get_FN124}
\alias{get_FN124}
\title{Get FN124 - Length Frequency Counts from FN_Portal API}
\usage{
get_FN124(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  uncount = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{uncount}{\itemize{
\item should the binned data be expanded to represent
indivual measurements.  A FN124 record with sizcnt=5 will be
repeated 5 times in the returned data frame.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for FN124 records from
the FN_Portal. FN124 records contain information about the length frequency
by species for each effort in a sample. For most gill
netting projects this corresponds to catches within a single panel
of a particular mesh size within a net set (gang). Group (GRP) is
occasionally included to further sub-divide the catch into user
defined groups that are usually specific to the project. This
function takes an optional filter list which can be used to return
record based on several attributes of the catch including species
or group code but also attributes of the effort, the sample or the
project(s) that the catches were made in.
}
\details{
See
http://10.167.37.157/fn_portal/api/v1/redoc/#operation/fn_124_list
for the full list of available filter keys (query parameters)
}
\examples{

fn124 <- get_FN124(list(lake = "ON", year = 2021, spc = "334"))

filters <- list(
  lake = "ON",
  protocol = "TWL",
  year = 2021,
  spc = c("061", "121"),
  sidep__lte = 40
)
fn124 <- get_FN124(filters)

LOA_IA21_TW1 <- get_FN124(list(prj_cd = "LOA_IA21_TW1"),
    uncount = TRUE)

}
\author{
Jeremy Holden \email{jeremy.holden@ontario.ca}
}
