% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN125.R
\name{get_FN125}
\alias{get_FN125}
\title{Get FN125 - Biological data from FN_Portal API}
\usage{
get_FN125(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN125
records. FN125 records contain the biological data collected from
individual fish sampled in assessment projects such as legth,
weight, sex, and maturity. For convience this end point also
returns data from child tables such as the 'preferred' age, and
lamprey wounds.  This function takes an optional
filter list which can be used to return records based on several
different biologial attributes (such as size, sex, or maturity),
but also of the species, or group code, or attributes of
the effort, the sample, or the project(s) that the
samples were collected in.
}
\details{
See
http://http://10.167.37.157/fn_portal/redoc/#operation/fn_123_list
for the full list of available filter keys (query parameters)
}
\examples{


fn125 <- get_FN125(list(lake='ON', year=2012, spc='334',gear='GL'))

filters <- list(lake='ER',
                protocol='TWL',
                spc_in=c('331', '334'),
                sidep_lte=20)
fn125 <- get_FN125(filters)

filters <- list(lake='SU',
           prj_cd__in=c('LSA_IA15_CIN','LSA_IA17_CIN'),
           eff='051',
           spc='091')
fn125 <- get_FN125(filters)

filters <- list(lake='HU', spc='076', grp='55')
fn125 <- get_FN125(filters)

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
