% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN125_Lamprey.R
\name{get_FN125_Lamprey}
\alias{get_FN125_Lamprey}
\title{Get FN125_Lam - Lamprey Wound data from FN_Portal API}
\usage{
get_FN125_Lamprey(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN125Lam
records. FN125Lam records contain information about the individual
lamprey wounds observed on a sampled fish.Historically, lamprey
wounds were reported as a single field (XLAM) in the FN125 table.
In the early 2000 the great lakes fishery community agreed to
capture lamprey wounding data in a more consistent fashion the
basin, using the conventions described in Ebener et al. 2006.  The
FN125Lam table captures data from individual lamprey wounds
collected using those conventions. A sampled fish with no
observed wound will have a single record in this table (with
lamijc value of 0), while fish with lamprey wounds, will have one
record for every observed wound. This function takes an optional
filter list which can be used to return records based on several
different attributes of the wound (wound type, degree of healing,
and wound size) as well as, attributes of the sampled
fish such as the species, or group code, or attributes of the
effort, the sample, or the project(s) that the samples were
collected in.
}
\details{
Use \code{show_filters("fn125lamprey")} to see the full list of
available filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/}
and filter by "fn125lamprey" for additional information.
}
\examples{

fn125Lam <- get_FN125_Lamprey(list(lake = "ON", spc = "081", grtp = "GL"))

fn125Lam <- get_FN125_Lamprey(list(
  lake = "SU", spc = "081", year__gte = 2015,
  lamijc_type = c("A1", "A2", "A3")
))

filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"),
  eff = "051",
  spc = "091"
)
fn125Lam <- get_FN125_Lamprey(filters)

filters <- list(lake = "HU", spc = "076", grp = "55")
fn125Lam <- get_FN125_Lamprey(filters)
fn125Lam <- get_FN125_Lamprey(filters, show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
