% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN125Lam.R
\name{get_FN125Lam}
\alias{get_FN125Lam}
\title{Get FN125Lam - Lamprey Wound data from FN_Portal API}
\usage{
get_FN125Lam(filter_list = list(), show_id = FALSE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN125Lam
records. FN125Lam records contain information about the individual
lamprey wounds observed on a sampled fish.  Historically, lamprey
wounds were reported as a single field (XLAM) in the FN125 table.
In the early 2000 the great lakes fishery community agreed to
capture lamprey wounding data in a more consistent fashion the
basin, using the conventions described in Ebener etal 2006.  The
FN125Lam table captures data from individual lamprey wounds
collected using those conventions.  A sampled fish with no
observed wound will have a single record in this table (with
lamijc value of 0), while fish with lamprey wounds, will have one
record for every observed wound.    This function takes an optional
filter list which can be used to return records based on several
different attributes of the wound (wound type, degree of healing,
and wound size) as well as, attributes of the sampled
fish such as the species, or group code, or attributes of the
effort, the sample, or the project(s) that the samples were
collected in.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn_125lam_list
for the full list of available filter keys (query parameters)
}
\examples{

fn125Lam <- get_FN125Lam(list(lake = "ON", spc = "081", gear = "GL"))

fn125Lam <- get_FN125Lam(list(
  lake = "SU", spc = "081", year__gte = 2015,
  lamijc_type = c("A1", "A2", "A3")
))

filters <- list(
  lake = "ER",
  protocol = "TWL",
  spc = c("331", "334"),
  year = 2010,
  sidep__lte = 20
)
fn125Lam <- get_FN125Lam(filters)

filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"),
  eff = "051",
  spc = "091"
)
fn125Lam <- get_FN125Lam(filters)

filters <- list(lake = "HU", spc = "076", grp = "55")
fn125Lam <- get_FN125Lam(filters)
fn125Lam <- get_FN125Lam(filters, show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
