% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn_portal_root.R
\name{get_fn_portal_root}
\alias{get_fn_portal_root}
\title{Internal function to return the root url for the FN_Portal api.}
\usage{
get_fn_portal_root()
}
\value{
string
}
\description{
This funciton is intended to be used by all of the functions that
call an api endpoint exposed by fn_portal.  If the api enpoint
changes in the future, (ie - the domain name changes or the
version changes) we just need to update this value.  Note that the
url string does \emph{NOT} include a trailing slash.
}
\details{
To use a local version of glfishr api during developement, set the
local environment variable to True:\preformatted{>  Sys.setenv(FN_PORTAL_DEV=TRUE)
}

By default, it is assumed that the api is being served on
localhost and port 8000, but these can be changed by setting the
environment variables FN_PORTAL_HOST, FN_PORTAL_PORT to the
appropriate host and port number.\preformatted{ >  Sys.setenv(FN_PORTAL_HOST='<some-ip-address>')
 >  Sys.setenv(FN_PORTAL_PORT=8080)
}
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
