% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gear_process_types.R
\name{get_gear_process_types}
\alias{get_gear_process_types}
\title{Get Gear Effort Process Types from the FN_Portal API}
\usage{
get_gear_process_types(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
Process type describes how the catch in a sample is handled -
either by net, by mesh size, by groups of panels, or individual
panels.  The available process types are constrained depending on
the gear type.  This function returns the known gear, effort and
process type combinations from the FN_Prortal api, which can then
be used to validee FN122 and FN123 records, or used to populate the
ProcessEffortGear table in the template database before submitting
to Process Validate.
}
\details{
Use ~show_filters("gear_effort_process_types")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/
and filter by "gear_effort_process_types" for additional information.
}
\examples{

ptype <- get_gear_process_types(list(gr = "GL10"))
ptype <- get_gear_process_types(list(gr = c("GL10", "GL21")))
ptype <- get_gear_process_types(list(gr__like = "GL1"))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
