% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid5s.R
\name{get_grid5s}
\alias{get_grid5s}
\title{Get 5-minute grids - A list of 5-minute grids for the Great Lakes}
\usage{
get_grid5s(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for 5-minute grids and returns
their number and the corresponding lake. It can fetch the entire table,
or it accepts a filter parameter for lake. The filter parameters 'page'
and 'page_size' are also accepted. No other filter
parameters are currently available for this endpoint.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/grid5s/
for the full list of 5-minute grids
}
\examples{

grid5s <- get_grid5s(list(page_size = 5000))
superior_grid5s <- get_grid5s(list(lake = "SU"))
grid5_slugs <- get_grid5s(list(page_size = 5000), show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
