% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gruse.R
\name{get_gruse}
\alias{get_gruse}
\title{Get gear use options - A list of gear use codes used in GLIS}
\usage{
get_gruse(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for gear use choices and returns
their labels, descriptions and whether they're in use. It fetches
the entire table of gear use codes - no other filter
parameters are currently available for this endpoint.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/gruse/
for the full list of gear use code options
}
\examples{

gruse <- get_gruse()
gruse_slugs <- get_gruse(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
