% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lakes.R
\name{get_lakes}
\alias{get_lakes}
\title{Get lakes - A list of lake names and abbreviations used in GLIS}
\usage{
get_lakes(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for lake names and returns
lake names and abbreviations. It is useful for fetching a table of
all lakes accepted by GLIS, but it also accepts the filter parameter
'lake' to specify one or more lakes (by abbreviation) to fetch.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/lakes/
for the full list of lakes
}
\examples{

lake_list <- get_lakes()
erie <- get_lakes(list(lake = "ER"))
upper_lakes <- get_lakes(list(lake = c("HU", "SU")))
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
