% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_management_units.R
\name{get_management_units}
\alias{get_management_units}
\title{Get management units from GLIS API}
\usage{
get_management_units(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{should the names of the dataframe be converted to
upper case?}
}
\value{
dataframe
}
\description{
Get management units from GLIS API
}
\examples{
get_management_units()
get_management_units(list(lake = "HU"))
get_management_units(list(lake = "HU", mu_type = "basin"))
}
