% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orientations.R
\name{get_orientations}
\alias{get_orientations}
\title{Get orientation options - A list of orientation codes used in GLIS}
\usage{
get_orientations(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for orientation choices and returns
their labels, descriptions and whether they're in use. It fetches
the entire table of orientation codes - no other filter
parameters are currently available for this endpoint.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/orient/
for the full list of orientation code options
}
\examples{

orientations <- get_orientations()
orientation_slugs <- get_orientations(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
