% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prj_leads.R
\name{get_prj_leads}
\alias{get_prj_leads}
\title{Get project leads - A list of project leads used in GLIS}
\usage{
get_prj_leads(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for project leads and returns
their names, usernames, email addresses, and whether they're active users
(current staff). It can fetch the entire table of active project leads,
or it accepts the filter parameter all=TRUE to return non-active project leads
too. No other filter parameters are currently available for this endpoint.
}
\details{
See
http://10.167.37.157/common/project_leads
for the full list of project leads
}
\examples{

prj_leads <- get_prj_leads()
all_prj_leads <- get_prj_leads(list(all=TRUE))
prj_lead_slugs <- get_prj_leads(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
