% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_leads.R
\name{get_project_leads}
\alias{get_project_leads}
\title{Get project leads - A list of project leads used in GLIS}
\usage{
get_project_leads(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for project leads and returns
their names and usernames. It can fetch the entire table of active project leads,
or it accepts the filter parameters to match all or part of a first name,
last name, or username.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/api/v1/swagger/
and filter by "project_leads" for a full list of available filter parameters.

See
https://intra.glis.mnr.gov.on.ca/common/project_leads/
for the full list of current and former staff that can be
entered as a project lead.
}
\examples{

project_leads <- get_project_leads()
all_project_leads <- get_project_leads(list(all = TRUE))
all_steves <- get_project_leads(list(first_name__like = "ste", all = TRUE))
project_lead_slugs <- get_project_leads(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
