% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_types.R
\name{get_project_types}
\alias{get_project_types}
\title{Get project type options - A list of project type codes used in GLIS}
\usage{
get_project_types(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for project type choices and returns
their labels, descriptions and whether they're in use. It fetches
the entire table of accepted project types - no other filter
parameters are currently available for this endpoint. Project type codes
represent the fifth and sixth digits of project code (PRJ_CD).
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/project_types/
for the full list of fisheries_office options
}
\examples{

fisheries_offices <- get_fisheries_offices()
fisheries_office_slugs <- get_fisheries_offices(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
