% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pt_abstract.R
\name{get_pt_abstract}
\alias{get_pt_abstract}
\title{get_pt_abstracts - Project abstract from FN_Portal API}
\usage{
get_pt_abstract(filter_list = list(), show_id = FALSE)
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for Project Tracker's abstract
records. Project tracker abstract records contain the custom text uploaded
by the project lead of an OMNR netting project.  The abstracts contain a
plain language summary of the project.  This api endpoint accepts a large number of filters
associated with the project including spatial filters for buffered
point, or region of interest.  Project specific filters include
project code(s), years, lakes, and project lead.  Use
'show_filter("projects")' to see the full list of available
filters.  This function returns a dataframe containing attributes
of project including project code, project name, start and end
date, and project lead.
}
\examples{
abstract <- get_pt_abstract(list(prj_cd = "LOA_IA17_GL1"))

}
\seealso{
\code{\link{show_filter}}
}
\author{
Jeremy Holden \email{jeremy.holden@ontario.ca}
}
