% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pt_abstracts.R
\name{get_pt_abstracts}
\alias{get_pt_abstracts}
\title{get_pt_abstracts - Project Tracker Abstracts}
\usage{
get_pt_abstracts(filter_list = list(), to_upper = TRUE, record_count = FALSE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the returned dataframe be
converted to upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for abstracts that have
been uploaded to project tracker. This api endpoint
accepts a large number of filters associated with the project or
report type. Project specific filters include project code(s),
years, lakes, and project lead. Use \code{show_filters("project_abstracts")}
to see the full list of available filters.
}
\examples{

abstracts <- get_pt_abstracts(list(
  lake = "ON", year__gte = 2012,
  year__lte = 2018
))

abstracts <- get_pt_abstracts(list(lake = "ER", year__gte = 2018))

filters <- list(lake = "SU", prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"))
abstracts <- get_pt_abstracts(filters)

abstracts <- get_pt_abstracts(list(lake = "HU", year__gte = 2018))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
