% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pt_associated_files.R
\name{get_pt_associated_files}
\alias{get_pt_associated_files}
\title{get_pt_associated_files - Project Tracker Associated Files}
\usage{
get_pt_associated_files(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the returned dataframe be
converted to upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for associated files that have
been uploaded to project tracker.  This api endpoint accepts a
large number of filters associated with the project.
Project specific filters include project code(s), years, lakes, and
project lead.  Use 'show_filters("associated_files")' to see the full list
of available filters.  This function returns a dataframe containing
attributes of the upload files, including project code, and
the path to the file on the server. It is often used in
conjunction with \code{\link[=fetch_pt_associated_files]{fetch_pt_associated_files()}} to actually download the
selected files to a target directory.
}
\examples{
\dontrun{
filters <- list(lake = "HU", year__gte = 2012, year__lte = 2018)
files <- get_pt_associated_files(filters)

filters <- list(lake = "HU", year__gte = 2012, prj_cd__like = "006")
files <- get_pt_associated_files(filters)

files <- get_pt_associated_files(list(lake = "ER"))

filters <- list(lake = "SU", prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"))
files <- get_pt_associated_files(filters)

files <- get_pt_associated_files(list(lake = "HU"))
}

}
\seealso{
\code{\link[=fetch_pt_associated_files]{fetch_pt_associated_files()}}
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
