% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pt_nr_milestones.R
\name{get_pt_nr_milestones}
\alias{get_pt_nr_milestones}
\title{get_pt_nr_milestones - Project Tracker Non-report Milestones}
\usage{
get_pt_nr_milestones(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the returned dataframe be
converted to upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for non-report milestones
in project tracker.  This api endpoint accepts a large number of filters
associated with the project or milestone type.  Project specific filters
include project code(s), years, lakes, and project lead.  Milestones
can also be filtered by their name.  Valid milestone types are: "Submitted",
"Approved", "Field Work Conducted", "Data Scrubbed", "Aging Complete",
"Data Merged", "Sign off", and "Cancelled" and can be passed in as a single
string or as charcter vector of one or more milestone types. Use
~show_filters("project_nr_milestones")~ to see the full list of available
filters. This function returns a dataframe containing attributes of the
report, including project code, project name, and the time at which a
milestone was completed.
}
\examples{

milestones <- get_pt_nr_milestones(list(
  lake = "ON", year__gte = 2012,
  year__lte = 2018
))

milestones <- get_pt_nr_milestones(list(
  lake = "HU", year__gte = 2012,
  prj_cd__like = "006", milestone = "Submitted"
))

milestones <- get_pt_nr_milestones(list(lake = "ER"))

filters <- list(lake = "SU", prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"))
milestones <- get_pt_nr_milestones(filters)
}
\seealso{
\code{\link[=fetch_pt_reports]{fetch_pt_reports()}}
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
