% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pt_points.R
\name{get_pt_points}
\alias{get_pt_points}
\title{get_pt_points - Project Tracker Points}
\usage{
get_pt_points(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the returned dataframe be
converted to upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for sampling points that
have been uploaded into project tracker.  This api endpoint accepts
a large number of filters associated with the project including
spatial filters for buffered point, or region of interest.  Project
specific filters include project code(s), years, lakes, and project
lead.  Use 'show_filter("points")' to see the full list of
available filters.  This function returns a data frame containing
attributes of project including project code, project project type,
dd_lat and dd_lon.
}
\examples{

points <- get_pt_points(list(lake = "ON", year__gte = 2012,
year__lte = 2018))

points <- get_pt_points(list(lake = "HU", year__gte = 2012,
prj_cd__like='006'))

filters <- list(lake = "SU", prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"))
points <- get_pt_points(filters)


}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
