% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pt_projects.R
\name{get_pt_projects}
\alias{get_pt_projects}
\title{get_pt_projects - Project Tracker Projects}
\usage{
get_pt_projects(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the returned dataframe be
converted to upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for projects in project
tracker.  This api endpoint accepts a large number of filters
associated with the project including spatial filters for buffered
point, or region of interest.  Project specific filters include
project code(s), years, lakes, and project lead.  Use
'show_filter("projects")' to see the full list of available
filters.  This function returns a dataframe containing attributes
of project including project code, project name, start and end
date, and project lead.
}
\examples{

projects <- get_pt_projects(list(
  lake = "ON", year__gte = 2012,
  year__lte = 2018
))

projects <- get_pt_projects(list(
  lake = "HU", year__gte = 2012,
  prj_cd__like = "006", report_type = "prtocol"
))

projects <- get_pt_projects(list(lake = "ER"))

filters <- list(lake = "SU", prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"))
projects <- get_pt_projects(filters)

projects <- get_pt_projects(list(lake = "HU"))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
