% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC012.R
\name{get_SC012}
\alias{get_SC012}
\title{Get SC012 - Sampling specs (species and group) from Creel Portal API}
\usage{
get_SC012(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for SC012 records.
SC012 records contain a list of anticipated species for a project and the
likely length and weight ranges for those species. These constraints are used
by ProcVal to check that biodata for fish caught during the project is
reasonable. The SC012 records contain information like minimum and maximum
TLEN, FLEN, RWT, and K values for each species. This function takes an
optional filter list which can be used to return records based on several
attributes of the project such as protocol, lake, project code, and species.
}
\details{
Use ~show_filters("sc012")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/
and filter by "sc012" for additional information.
}
\examples{

sc012 <- get_SC012(list(lake = "HU"))

filters <- list(prj_cd = c("LHA_SC17_DR7", "LHA_IA22_100"))
fn012 <- get_FN012(filters)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
