% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC012_Lake.R
\name{get_SC012_Lake}
\alias{get_SC012_Lake}
\title{Get FN012_Lake - Sampling specs (species and group) for a specific lake from Creel_Portal API}
\usage{
get_SC012_Lake(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for FN012 Lake List records.
These records are the expected range of biodata values
(e.g. FLEN, TLEN, RWT, K) for the typical species caught during a creel
survey in a given lake. These constraints are used by ProcVal to check
that biodata for fish caught during the project is reasonable.
The lake MUST be specified for retrieving records using this function.
This is a useful function for building a template database for a new
creel project or for adding an FN012 table to an existing project.
}
\details{
Use ~show_filters("sc012_protocol")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/
and filter by "sc012_protocol" for additional information.
}
\examples{

sc012_lake <- get_SC012_Lake(list(lake = "ON"))

sc012_lake <- get_SC012_Lake(list(lake = "ER", biosam = "1"))

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
