% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC012_Protocol.R
\name{get_SC012_Protocol}
\alias{get_SC012_Protocol}
\title{Get SC012_Protocol - Creel Sampling specs (species and group) for a
specific lake and from the Creel_Portal API}
\usage{
get_SC012_Protocol(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for SC012 Protocol List
records. These records are the expected range of biodata values
(e.g. FLEN, TLEN, RWT, K) for the typical species sampled in a
creel conducted in a given lake. These constraints are used
by ProcVal to check that biodata for fish sampled during the creel is
reasonable. This is a useful function for building
a template database for a new project or for adding an SC012 table to
an existing project.
}
\details{
Use \code{show_filters("sc012_protocol")} to see the full list of available filter
keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/sc_portal/api/v1/swagger/}
and filter by "sc012_protocol" for additional information.
}
\examples{

sc012_protocol <- get_SC012_Protocol(list(lake = "ON"))

sc012_protocol <- get_SC012_Protocol(list(lake = c("HU", "ER"), protocol = "BSM"))

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
