% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC022.R
\name{get_SC022}
\alias{get_SC022}
\title{Get SC022 - Season Strata Data from Creel_Portal API}
\usage{
get_SC022(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for SC022
records. SC022 records contain information about the defined season
associated with a creel.
}
\details{
Use \code{show_filters("sc022")} to see the full list of available
filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/} and
filter by "sc022" for additional information.
}
\examples{

sc022 <- get_SC022(list(lake = "ON", year = 2012))

sc022 <- get_SC022(list(prj_cd = "LHA_SC11_053"))
sc022 <- get_SC022(list(prj_cd = "LHA_SC11_053"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
