% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC023.R
\name{get_SC023}
\alias{get_SC023}
\title{Get SC023 - Day Type Data Creel_Portal API}
\usage{
get_SC023(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for SC023 records. SC023
records contain information about the day types considered in a
creel and the codes used to identify those days. Days are usually
identified as either weekday, or weekend days. (the SC025 end point
is used to identify exception dates - weekdays that should be
treated like a weekend).
}
\details{
Use ~show_filters("sc023")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/
and filter by "sc023" for additional information.
}
\examples{

sc023 <- get_SC023(list(lake = "ON", year = 2012))

sc023 <- get_SC023(list(prj_cd = "LHA_IA19_812"))
sc023 <- get_SC023(list(prj_cd = "LHA_IA19_812"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
