% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC025.R
\name{get_SC025}
\alias{get_SC025}
\title{Get SC025 - Exception Dates from Creel_Portal API}
\usage{
get_SC025(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for SC025 records. SC025
records identify exception dates - weekdays that should be treated
like a weekend.
}
\details{
Use \code{show_filters("sc025")} to see the full list of available
filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/} and
filter by "sc025" for additional information.
}
\examples{

sc025 <- get_SC025(list(lake = "ON", year = 2012))

sc025 <- get_SC025(list(prj_cd = "LOA_SC12_CRS"))
sc025 <- get_SC025(list(prj_cd = "LOA_SC12_CRS"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
