% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC026.R
\name{get_SC026}
\alias{get_SC026}
\title{Get SC026 - Spatial Strata Data Creel_Portal API}
\usage{
get_SC026(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for SC026 records. SC026
records contain information about the spatial strata associated
with a project.
}
\details{
Use ~show_filters("sc026")~ to see the full list of available filter
keys (query parameters)
}
\examples{

sc026 <- get_SC026(list(lake = "ON", year = 2012))

sc026 <- get_SC026(list(prj_cd = "LOA_SC12_002"))
sc026 <- get_SC026(list(prj_cd = "LOA_SC12_002"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
