% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC028.R
\name{get_SC028}
\alias{get_SC028}
\title{Get SC028 - Fishing Modes from Creel_Portal API}
\usage{
get_SC028(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for SC028 records. SC028
records contain information about the fishing modes included in a
creel survey.  Fishing mode typically describes the gear, or method
of fishing.
}
\details{
Use ~show_filters("sc028")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/
and filter by "sc028" for additional information.
}
\examples{

sc028 <- get_SC028(list(lake = "ON", year = 2012))

sc028 <- get_SC028(list(prj_cd = "LOA_SC12_002"))
sc028 <- get_SC028(list(prj_cd = "LOA_SC12_002"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
