% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC121.R
\name{get_SC121}
\alias{get_SC121}
\title{Get SC121 - Anlger Interviews from Creel_Portal API}
\usage{
get_SC121(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for SC121 records. SC121
records contain information about interviews with anglers conducted
during a creel survey.  The SC121 records contain information like
the number of anglers in there party, when the started fishing and
often answers to creel specific optional questions. This function
takes an optional filter list which can be used to return record
based on several attributes of the interview.
}
\details{
Use \code{show_filters("sc121")} to see the full list of available
filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/} and
filter by "sc121" for additional information.
}
\examples{

sc121 <- get_SC121(list(lake = "HU", year = 2009))

sc121 <- get_SC121(list(lake = "HU", prj_cd__like = "_001", year = 2009))
sc121 <- get_SC121(list(lake = "HU", prj_cd__like = "_001", year = 2009), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
