% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC123.R
\name{get_SC123}
\alias{get_SC123}
\title{Get SC123 - Catch Counts  from Creel_Portal API}
\usage{
get_SC123(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for SC123 records from
the SC_Portal. SC123 records contain information about catch counts
by species and the target species for each creel interview..  This
function takes an optional filter list which can be used to return
record based on several attributes of the catch including species,
or group code but also attributes of the interview or the
creel that the catches were made in.
}
\details{
Use ~show_filters("sc123")~ to see the full list of available filter
keys (query parameters)
}
\examples{

sc123 <- get_SC123(list(lake = "HU", year = 2000, spc = "078"))

filters <- list(
  lake = "ER",
  year = 2010,
  spc = c("331", "334")
)
sc123 <- get_SC123(filters)


filters <- list(lake = "HU", spc = "076", sek=TRUE)
sc123 <- get_SC123(filters)

sc123 <- get_SC123(list(prj_cd = "LHA_SC09_033"))
sc123 <- get_SC123(list(prj_cd = "LHA_SC09_033"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
