% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC124.R
\name{get_SC124}
\alias{get_SC124}
\title{Get SC124 - Length Frequency Counts from Creel_Portal API}
\usage{
get_SC124(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  uncount = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the data-frame be converted to
upper case?
}}

\item{uncount}{\itemize{
\item should the binned data be expanded to represent
individual measurements.  A SC124 record with sizcnt=5 will be
repeated 5 times in the returned data frame.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for SC124 records from the
Creel_Portal. SC124 records contain information about the length
frequency by species associated with an angler interview. Group
(GRP) is occasionally included to further sub-divide the catch into
user defined groups that are usually specific to the project. This
function takes an optional filter list which can be used to return
record based on several attributes of the catch including species
or group code but also attributes of the effort, the sample or the
project(s) that the catches were made in.
}
\details{
Use \code{show_filters("sc124")} to see the full list of available
filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/} and
filter by "sc124" for additional information.
}
\examples{

sc124 <- get_SC124(list(lake = "ON", year = 2021, spc = "334"))

filters <- list(
  lake = "ON",
  year = 2021,
  spc = c("061", "121")
)
sc124 <- get_SC124(filters)

sc124 <- get_SC124(list(prj_cd = "LOA_SC19_002"),
  uncount = TRUE
)
}
\author{
Jeremy Holden \email{jeremy.holden@ontario.ca}
}
