% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC125.R
\name{get_SC125}
\alias{get_SC125}
\title{Get SC125 - Biological data from Creel_Portal API}
\usage{
get_SC125(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE,
  add_year_col = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns as
part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}

\item{add_year_col}{\itemize{
\item should a 'year' column be added to the
returned dataframe?  This argument is ignored if the data frame
does not contain a 'prj_cd' column.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for SC125 records. SC125
records contain the biological data collected from individual fish
sampled in creel surveys such as length, weight, sex, and maturity.
This function takes an optional filter list which can be used to
return records based on several different biological attributes
(such as size, sex, or maturity), but also of the species, or group
code, or attributes of the effort, the sample, or the project(s)
that the samples were collected in.
}
\details{
Use \code{show_filters("sc125")} to see the full list of available
filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/} and
filter by "sc125" for additional information.
}
\examples{


sc125 <- get_SC125(list(lake = "ON", year = 2012, spc = "334", gr = "GL"))

filters <- list(
  lake = "ER",
  spc_in = c("331", "334"),
  year__gte = 2014, year__gle = 2016
)
sc125 <- get_SC125(filters)

filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"),
  eff = "051",
  spc = "091"
)
sc125 <- get_SC125(filters)

filters <- list(lake = "HU", spc = "076", grp = "55", year__gte = 2016)
sc125 <- get_SC125(filters)

sc125 <- get_SC125(list(prj_cd = "LHA_IA19_812"))
sc125 <- get_SC125(list(prj_cd = "LHA_IA19_812"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
