% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC126.R
\name{get_SC126}
\alias{get_SC126}
\title{Get SC126 - Diet data from Creel_Portal API}
\usage{
get_SC126(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for SC126 records. SC126
records contain the counts of identifiable items in found in the
stomachs of sampled fish.  This function takes an optional filter
list which can be used to return records based on several different
attributes of the diet item (taxon, taxon_like), as well as,
attributes of the sampled fish such as the species, or group code,
or attributes of interview, or the creel(s) that the samples were
collected in.
}
\details{
Use \code{show_filters("sc126")} to see the full list of available
filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/} and
filter by "sc126" for additional information.
}
\examples{

sc126 <- get_SC126(list(lake = "HU", year = 2000, spc = "334"))

filters <- list(
  lake = "ER",
  protocol = "TWL",
  spc = c("331", "334"),
  sidep0__lte = 20
)
sc126 <- get_SC126(filters)

filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"),
  eff = "051",
  spc = "091"
)
sc126 <- get_SC126(filters)

filters <- list(lake = "HU", spc = "076", grp = "55")
sc126 <- get_SC126(filters)
sc126 <- get_SC126(filters, show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
