% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sc_portal_root.R
\name{get_sc_portal_root}
\alias{get_sc_portal_root}
\title{Internal function to url to the SC_Portal api.}
\usage{
get_sc_portal_root()
}
\value{
string
}
\description{
This function is intended to be used by all of the functions that
call an api endpoint exposed by sc_portal. Each of the portals
(tfat, sc_portal, sc_portal etc.)  all have a slightly different
access point, and each will have a similar function. If the api
endpoint for that portal changes in the future, (i.e.  the version
changes) we just need to update the value in this file and all of
the dependent functions will use it.  Note that the url string does
\emph{NOT} include a trailing slash.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
