% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get Species list data from FN_Portal API}
\usage{
get_species(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for Species list
available in Fn_portal. The species list contain all of the fields
in common lookup table, but the fields are limited to spc,
spc_nmco, and spc_nmsc.  This
function takes an optional filter list which can be used to return
species based on part of their common or scientific name.
}
\details{
See
http://http://10.167.37.157/fn_portal/redoc/#operation/species_list
for the full list of available filter keys (query parameters)
}
\examples{

species <- get_species()
trout <- get_species(list(spc_nmco_like='trout'))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
