% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Stream_Dimensions.R
\name{get_Stream_Dimensions}
\alias{get_Stream_Dimensions}
\title{Get Stream Dimensions data from FN_Portal API}
\usage{
get_Stream_Dimensions(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'slug' field is
hidden from the data frame. To return this field as part of the
data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for Stream_Dimensions
records. Stream_Dimensions records contain information about
streams and rivers in which electrofishing transects take place,
including the distance from the start of the transect, the distance
across the channel, and stream depth, width, and velocity. Other
relevant details for the SUBSPACE are found in the FN121 table.
This function takes an optional filter list which can be used to
return records based on attributes of the project including project
code, or part of the project code, lake, first year, last year,
protocol, etc.
}
\details{
Use \code{show_filters("stream_dimensions")} to see the full list
of available filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/}
and filter by "stream_dimensions" for additional information.
}
\examples{
\dontrun{
show_filters("stream_dimensions")
}
# TODO: add more examples when more data is uploaded
stream <- get_Stream_Dimensions(list(lake = "ER"))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
