% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_filters.R
\name{get_swagger_payload}
\alias{get_swagger_payload}
\title{Connect to the api endpoint and return swagger documentation}
\usage{
get_swagger_payload(api_app)
}
\arguments{
\item{api_app}{\itemize{
\item the name of the api application to fetch the
filters from.
}}
}
\value{
list
}
\description{
This function connects to the openapi/swagger endpoint specified by
the api_api argument and convertes the json response into an r list
(payload) that can be parsed by other functions to extract
information about the available filters or data fields contained in
the response from each endpoint.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
