% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_filters.R
\name{get_swagger_url}
\alias{get_swagger_url}
\title{Build the url to fetch the swagger data for a give app}
\usage{
get_swagger_url(
  api_app = c("common", "project_tracker", "fn_portal", "creels", "tfat")
)
}
\arguments{
\item{api_app}{\itemize{
\item the name of the api application to fetch the
filters infromation from.
}}
}
\value{
character vector
}
\description{
An internal  helper function that combines the applicaiton name
with the current domain to build the url for the swagger
documentation for that app.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
