% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_agencies.R
\name{get_tag_agencies}
\alias{get_tag_agencies}
\title{Get tagging agency options - A list of tagging agency codes used in GLIS}
\usage{
get_tag_agencies(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for tagging agency choices and returns
their labels, descriptions and whether they're in use. It fetches
the entire table of tagging agency codes - no other filter
parameters are currently available for this endpoint.
Tagging agency is the third and fourth characters of TAGDOC.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/tagging_agencies/
for the full list of tagging agency code options
}
\examples{

orientations <- get_orientations()
orientation_slugs <- get_orientations(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
