% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_attr_values.R
\name{get_tag_attr_values}
\alias{get_tag_attr_values}
\title{Get the list of values for tag attributes from the TFAT api}
\usage{
get_tag_attr_values(
  what = c("tag_origin", "tag_colour", "tag_type", "tag_position", "tagstat")
)
}
\arguments{
\item{what}{string}
}
\value{
dataframe
}
\description{
This function accesses the tfat lookup api endpoint and extracts the
available values for  tag colour, placement, type, status (applied/on
capture), or tag origin (i.e. - tagging agency). The tag data is returned as
a data frame with the columns 'value' and 'label'.
}
\examples{

tag_origins <- get_tag_attr_values("tag_origin")

tag_colours <- get_tag_attr_values("tag_colour")

tag_tag_types <- get_tag_attr_values("tag_type")

tag_positions <- get_tag_attr_values("tag_position")

tagstats <- get_tag_attr_values("tagstat")
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
