% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_colours.R
\name{get_tag_colours}
\alias{get_tag_colours}
\title{Get tag colour options - A list of tag colour codes used in GLIS}
\usage{
get_tag_colours(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for tag colour choices and returns
their labels, descriptions and whether they're in use. It fetches
the entire table of accepted tag colours - no other filter
parameters are currently available for this endpoint. Tag colour is the
fifth character of TAGDOC.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/tag_colours/
for the full list of tag colour options
}
\examples{

tag_colours <- get_tag_colours()
tag_colour_slugs <- get_tag_colours(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
