% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_types.R
\name{get_tag_types}
\alias{get_tag_types}
\title{Get tag types - A list of tag types used in GLIS}
\usage{
get_tag_types(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for tag type choices and returns
their labels, descriptions and whether they're in use. It can fetch
the entire table of accepted tag types, or it accepts the filter parameter
all=TRUE to return depreciated tag type choices too. No other filter
parameters are currently available for this endpoint. Tag types are
first character of TAGDOC.
}
\details{
See
http://10.167.37.157/common/tag_types
for the full list of tag types
}
\examples{

tag_types <- get_tag_types()
all_tag_types <- get_tag_types(list(all = TRUE))
tag_slugs <- get_tag_types(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
