% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxon_table.R
\name{get_taxon_table}
\alias{get_taxon_table}
\title{Get taxon table - A list of fish and non-fish species and their ITIS taxon code}
\usage{
get_taxon_table(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for taxon records of fish and
non-fish species. These records include the ITIS code, common name, scientific
name, taxonomic rank, vert/invert classification, and HHFAU code (if applicable).
Different taxonomic ranks allow selection of a taxon at different levels of
specificity (e.g. 'Turtles/Testudines' vs. 'Eastern painted turtle').
}
\details{
Current filter parameters are 'taxon', 'itiscode', 'taxon_name', 'taxon_label',
'taxonomic_rank', 'vertinvert', and 'omnr_provincial_code'.

See
http://10.167.37.157/common/taxon
for the full list of taxon codes
}
\examples{

taxa_list <- get_taxon_table()
order_list <- get_taxon_table(list(taxonomic_rank = "order"))
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
