% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfat_projects.R
\name{get_tfat_projects}
\alias{get_tfat_projects}
\title{Get TFAT Tagging Projects}
\usage{
get_tfat_projects(filter_list = list(), record_count = FALSE)
}
\arguments{
\item{filter_list}{list}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the TFAT api endpoint for tagging projects - OMNR
field projects that either recaptured or applied at least one fish tag
(excluding CWTs).  The TFAT project endpoint contains only basic information
about a project such as project code, project name, project lead, start and
and date.  For more detailed information about projects, please see Project
Tracker.
}
\details{
This function takes an optional filter list which can be used to return
records based on several attributes of the projects such as lake or year they
were run, attributes of the tagged fish such as species or size, the tag
attributes such as colour, placement, tag type, or tag status (either tagged
or recovered).
}
\examples{

# projects that applied Lake Huron between 2012 and 2018
projects <- get_tfat_projects(list(
  lake = "HU", year__gte = 2012, year__lte =
    2018, tagstat = "A"
))

# projects that recovered at least one tagged walleye or musky
projects <- get_tfat_projects(list(spc = c("334", "132"), tagstat = "C"))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
