% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfat_recoveries.R
\name{get_tfat_recoveries}
\alias{get_tfat_recoveries}
\title{Get TFAT Tag Recoveries}
\usage{
get_tfat_recoveries(filter_list = list(), record_count = FALSE)
}
\arguments{
\item{filter_list}{list}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the TFAT api endpoint to for tag recoveries
\itemize{
\item tags reported by anglers, the general public, or other agencies.
Tag Recovery events include information about the tag, the tagged
fish, and where the recovery occurred.
}
}
\details{
This function takes an optional filter list which can be used to
return records based on several attributes of the tagged fish
(species, or size), the tag attributes (colour, placement, tag
type) or the recovery event (lake or recovery date)

See \code{get_tag_attr_values()} to get a list of values that can
be use to filter results based on tag attributes such as colour or
tag type.
}
\examples{

# tags recovered in Lake Huron between 2012 and 2018
recoveries <- get_tfat_recoveries(list(
  lake = "HU", year__gte = 2012,
  year__lte = 2018
))

# yellow or orange tags recovered in Lake Superior
recoveries <- get_tfat_recoveries(list(lake = "SU", tag__colour = c("2,5")))

# tags recovered from walleye or musky
recoveries <- get_tfat_recoveries(list(spc = c("334", "132")))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
