% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfat_anglers.R
\name{get_tfat_reporters}
\alias{get_tfat_reporters}
\title{Get TFAT Tag Reporters}
\usage{
get_tfat_reporters(filter_list = list(), record_count = FALSE)
}
\arguments{
\item{filter_list}{list}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the TFAT api endpoint to for tag reporters - basic
information about entities who have reported tags. Most often these are
anglers, the general public, or other agencies.  This api endpoint returns
basic information about tag reporters including first and last name,
affiliation, and contact information. This api endpoint is only available to
authenticated and authorized users.
}
\details{
This function takes an optional filter list which can be used to return
records based on several attributes of the reporter (first and last name
'like'),  the reporting event (reporting date, reporting method, or follow-up
requested), or the tag recovery event such as the species or size of fish or
the tag attributes such as colour, placement, or tag type.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
