% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfat_reports.R
\name{get_tfat_reports}
\alias{get_tfat_reports}
\title{Get TFAT Tag Recovery Reports}
\usage{
get_tfat_reports(filter_list = list(), record_count = FALSE)
}
\arguments{
\item{filter_list}{list}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the TFAT api endpoint for tag reports - tag reports
are anglers, the general public, or other agencies.  A tag report can have
one or many associated tags (i.e. tag recoveries).  Every tag recovery has
only one tag report. Tag reports include the reporting date and method, and
whether or not a follow-up letter  was requested.
}
\details{
This function takes an optional filter list which can be used to return
records based on several attributes of the reporting event (reporting date,
reporting method, or follow-up requested), the tagged fish such as species or
size, or the tag attributes such as colour, placement, or tag type.
}
\examples{

# tags recovered in Lake Huron between 2012 and 2018
reports <- get_tfat_reports(list(
  lake = "ON", year__gte = 2012, year__lte =
    2018
))

# yellow or orange tags recovered in Lake Superior
reports <- get_tfat_reports(list(lake = "SU", tag__colour = c("2,5")))

# tags recovered from walleye or musky
reports <- get_tfat_reports(list(spc = c("334", "132")))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
