% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfat_root.R
\name{get_tfat_root}
\alias{get_tfat_root}
\title{Internal function to url to the tfat api.}
\usage{
get_tfat_root()
}
\value{
string
}
\description{
This function is intended to be used by all of the functions that call an api
endpoint exposed by tfat. Each of the portals (tfat, sc_portal, fn_portal
etc.) all have a slightly different access points, and each will have a
similar function. If the api endpoint for that portal changes in the future,
(i.e. the  version changes) we just need to update the value in this file and
all of the dependent functions will use it.  Note that the url string does
\emph{NOT} include a trailing slash.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
