% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissues.R
\name{get_tissues}
\alias{get_tissues}
\title{Get tissues - A list of tissue types used in GLIS}
\usage{
get_tissues(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for tissue types and returns
their names, descriptions and whether they're in use. It can fetch
the entire table of accepted tissue codes, or it accepts filter parameter
all=TRUE to return depreciated tissue codes too. No other filter
parameters are currently available for this endpoint.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/tissues/
for the full list of tissues
}
\examples{

tissues <- get_tissues()
all_tissues <- get_tissues(list(all = TRUE))
tissue_slugs <- get_tissues(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
