% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get GLIS API token}
\usage{
get_token()
}
\value{
string
}
\description{
A helper function to fetch a token from the GLIS API. The token is
added to the GET request header. A new token is needed every 24 hours
(or when the existing token is cleared from the Global Environment).
}
