% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trg_table_names.R
\name{get_trg_table_names}
\alias{get_trg_table_names}
\title{Get field names for target table}
\usage{
get_trg_table_names(trg_db, table)
}
\arguments{
\item{trg_db}{\itemize{
\item path to our target accdb file
}}

\item{table}{\itemize{
\item the table name in the target databae to query against
}}
}
\value{
character vector
}
\description{
This function connects to at target database and execute a simple
sql statement that will return an empty record set that contains
nothing but the names of the columns in the target table.  This
list of field names can be used to modify or map existing data
frames to match the target table so that subsequent append queries
have matching schema.
}
\seealso{
sync_fields
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
