% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessels.R
\name{get_vessels}
\alias{get_vessels}
\title{Get vessels - A list of vessel names used in GLIS}
\usage{
get_vessels(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for vessels and returns
their names, descriptions and whether they're in use. It can fetch
the entire table of active vessels, or it accepts filter parameter
all=TRUE to return depreciated vessels too. No other filter
parameters are currently available for this endpoint.
}
\details{
See
http://10.167.37.157/common/vessels
for the full list of vessels
}
\examples{

vessel_list <- get_vessels()
all_vessels <- get_vessels(list(all = TRUE))
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
