% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glis_csv.R
\name{glis_csv}
\alias{glis_csv}
\title{Fetch and Read csv files from glis}
\usage{
glis_csv(filepath)
}
\arguments{
\item{filepath}{string}
}
\value{
dataframe
}
\description{
This function takes a path from a csv file that has been uploaded
GLIS and reads it into a data frame.  To read files directly from
glis using R, requires a user token.  If a token is available in
the current R session it will be used, otherwise a new token will
be generated.  The function's argument is a string that represents
the last part of the download url for the target file.  This is
consistent with the FILE_PATH attribute of records returned by the
~get_pt_associated_files()~ function.  Alternatively, the complete
file path including the domain and complete url is also valid.
}
\examples{

# it is possible to fetch the list of files using the glfishr
# function:
files <- get_pt_associated_files(list(prj_cd = "LHA_IA24_116"))
# you might have ot filter the path objects to ensure they are the correct csv's

csv_url <- files[1, ]$FILE_PATH
xlymph_data <- glis_csv(csv_url)
head(xlymph_data)
# or
csv_url <- files[2, ]$FILE_PATH
file2 <- glis_csv(csv_url)
head(file2)

#----------------------------#'
# OR

# first right on the link provided in the details page and select
# "Copy link address" from the pop-up menu and paste it here:
csv_url <- paste0(
  "https://intra.glis.mnr.gov.on.ca/project_tracker/",
  "serve_file/associated_files/LHA_IA24_116/LHA_IA24_116-XLYMPH.csv"
)

# then pass that url to the  glis_csv() function
csv_data <- glis_csv(csv_url)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
