% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{insert_gps_tracks}
\alias{insert_gps_tracks}
\title{Insert FN121_GPS_Tracks}
\usage{
insert_gps_tracks(dbase, data)
}
\arguments{
\item{dbase}{\itemize{
\item the absolute path to the target template database
}}

\item{data}{the data to be inserted into the FN121_GPS_tracks table}
}
\value{
the status of the rodbc connection
}
\description{
Insert data into the FN121_GPS_Tracks Table
}
\details{
The gps tracks table contains the field "TIMESTAMP" which is a key
word in MS access but is not being escaped properly by the query
constructed by the RODBC packages. This hack, inserts the gps data
into a temporary table with an alternative field name
"TIME_STAMP".  this temporary table is then inserted into
FN121_GPS_Tracks using a correctly formated sql statement, and
then deletes the tempoary table from the database.

The odbc package appears to work in this case, but would require a
rewrite of both glfishr and glisDbTools to ensure consistent dependenies.
}
\author{
R. Adam Cottrill
}
