% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mu_slug.R
\name{parse_mu_slug}
\alias{parse_mu_slug}
\title{Parse Management Unit Slug}
\usage{
parse_mu_slug(mu_slug, toupper = TRUE)
}
\value{
string
}
\description{
Using the mu_type filter to add management units to the FN121 table results in the
field being populated with the management unit slug. Use \code{parse_mu_slug} to replace
the slug with the proper name of the management unit. The default for parsed slugs
is all upper case - use toupper=FALSE for title case (first letter of each word
capitalized).
}
\examples{
huron_offshore_19 <- get_FN121(list(lake = "HU", protocol = "OSIA", year = 2019, mu_type = "basin"))
huron_offshore_19$MANAGEMENT_UNIT <- parse_mu_slug(huron_offshore_19$MANAGEMENT_UNIT)

}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
