% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_filters.R
\name{parse_swagger_payload}
\alias{parse_swagger_payload}
\title{Convert swagger response into named list}
\usage{
parse_swagger_payload(payload)
}
\arguments{
\item{payload}{\itemize{
\item list parsed from swagger api response
}}
}
\value{
list
}
\description{
the response from the swagger endpoint include a large number of
attributes and details that are not need for glfishr. this
function takes the named list created from the json response and
extracts the name and description of each GET parameter that can
be used as filters.
}
\details{
Given the complexity of the payload, this function is not
currently tested

TODO: add option to parse schemas from payload
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
