% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_dataframe.R
\name{pluck_dataframe}
\alias{pluck_dataframe}
\title{Pluck a data-frame from a named list of key-value pairs}
\usage{
pluck_dataframe(payload, what = "tag_type")
}
\arguments{
\item{payload}{list}

\item{what}{string}
}
\value{
dataframe
}
\description{
This is a helper function that is used to extract or pluck a
data-frame from a named list of key-values pairs.  Api endpoints
for lookup-values are often in the form of key-value pairs. This
function selects the named element from list and returns it as a
data-frame with the columns 'value' and 'label'. This function is
not expected to be called directly, but is used by other
convenience functions (\code{get_tag_colours()},
\code{get_tag_types()}).
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
