% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{populate_db}
\alias{populate_db}
\title{Populate target database}
\usage{
populate_db(target, data, verbose, verbose_sqlsave)
}
\arguments{
\item{target}{\itemize{
\item the path to the target accdb file
}}

\item{data}{\itemize{
\item a named list containing the data to insert the
target db
}}

\item{verbose}{\itemize{
\item should the table and record count be reported
}}

\item{verbose_sqlsave}{\itemize{
\item Boolean. Default=FALSE. This value is
passed through to RODBC::sqlsave.  Set to True to see all of
the sql statements passed to target database.  Useful for
debugging.
}}
}
\description{
given a list containing our data elements and a target database,
iterate over the list and insert the data into the corresponding
table in the target database.  The names in the list must
correspond with the tables in the target database, and it must be
possible to insert the tables in alphabetical order. (which works
for FN-2, but may not always be the case)
}
\seealso{
append_data, populate_template_assessment
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
