% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_fn012.R
\name{populate_fn012}
\alias{populate_fn012}
\title{Populate the FN012-Sampling Specs Table}
\usage{
populate_fn012(
  filters,
  glis_data,
  prune_fn012,
  source = c("assessment", "creel")
)
}
\arguments{
\item{filters}{\itemize{
\item list of filters used to select projects
}}

\item{glis_data}{\itemize{
\item the named list data fetched from the glis api
}}

\item{prune_fn012}{\itemize{
\item boolean - should unused FN012 records be
removed from the table?
}}

\item{source}{\itemize{
\item 'assessment' or 'creel'
}}
}
\value{
dataframe
}
\description{
This function populates the FN012 table for project selected by
the provided filters.  Attributes of the FN011, FN123, FN124 and
FN125 elements in the data list are used to augment rows and their
attributes in instances where those rows don't already exist.  If
\code{\link{prune_fn012}} is TRUE only records with at least one
corresponding record in the FN123 table are returned.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
