% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{populate_gept}
\alias{populate_gept}
\title{Populate Gear-Effort-Proecss-Type from FN028 and FN121 tables}
\usage{
populate_gept(fn028, fn121)
}
\arguments{
\item{fn028}{\itemize{
\item FN028-mode dataframe with coulmn 'GR'
}}

\item{fn121}{\itemize{
\item FN121 - net set dataframe with column 'PROCESS_TYPE'
}}
}
\value{
dataframe
}
\description{
The Gear-Effort-Process-Type table in the GLIS assessment table is
used to constrain the efforts that can appear in the FN122 table,
depending on the process type reported in the FN121 table and
associated gear type.  This function creates the records in that
table by fetching all of the known process types for a given gear
and then filter those results to include only those used in the
provided dataframe.
}
\author{
Arthur Bonsall \email{arthur.bonsall@ontario.ca}
}
