% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{prepare_payload}
\alias{prepare_payload}
\title{Perform transformations to api payload before return data frame}
\usage{
prepare_payload(payload, show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{payload}{dataframe}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function takes the dataframe returned from the api call and
preforms any required transformations before returning it.
Currently the fuction optionally removes is and slug from the data
frame (if they exist), and transforms all ofhte field names to
uppercase so that they match names that have been traditionally
used in FISHNET-2.  Other transfomations or modifications could be
added in the future.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
