% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{prune_unused_fn012}
\alias{prune_unused_fn012}
\title{Remove FN012 records without matching FN123 records}
\usage{
prune_unused_fn012(fn012, fn123)
}
\arguments{
\item{fn012}{\itemize{
\item dataframe with columns PRJ_CD, SPC, and GRP
containing sampling specifications
}}

\item{fn123}{\itemize{
\item dataframe with columns PRJ_CD, SPC, and GRP
containing catch count information.
}}
}
\value{
dataframe
}
\description{
A helper function used by populate_fn012 to removed any sampling
specification s for species-groups that were not encountered in a
project.
}
\seealso{
get_or_augment_fn012
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
