% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfishr_utils.R
\name{sam_slug}
\alias{sam_slug}
\title{Get FN121 Sam Slug}
\usage{
sam_slug(slug)
}
\arguments{
\item{slug}{\itemize{
\item string.  Often a human readable, unique idnetifier
}}
}
\value{
string
}
\description{
Get FN121.SAM slug
}
\details{
This function accepts a slug and returns a string that corresponds
to the net-set slug.  Joining records to the FN121 table is so
common, that a designated function is warrented, rather than
recursively calling trim_slug() or manually pasting prj_cd and sam
together.  The regular expression takes any slug and returns the
part of that string that corresponds to the project code and
sample number.
}
\author{
R. Adam Cottrill
}
