% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_filters.R
\name{show_filters}
\alias{show_filters}
\title{Show filters available on the provided api endpoint}
\usage{
show_filters(endpoint = "", filter_like = "")
}
\arguments{
\item{endpoint}{\itemize{
\item the name of the api endpoint
}}

\item{filter_like}{\itemize{
\item a partial string to match filter against
}}
}
\description{
Most of the API endpoints have a large number of
available filters. This function takes an API endpoint name and
prints out all of the available filters. If 'filter_like' is
provided, only those filters that contain the provided string are
printed. If no endpoint name is provided, a list of available
endpoints will be displayed.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/project_tracker/api/v1/swagger/,
https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/,
https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/, and
https://intra.glis.mnr.gov.on.ca/common/api/v1/swagger/
for an alternative way to view the full list of available filter keys (query parameters)
}
\examples{
\dontrun{
show_filters("fn125")
show_filters("fn125", filter_like = "prj")
show_filters("sc121", "mode")
show_filters("species")
show_filters("sc121", "cats")
# show_filters("foo")
}
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
