% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{sync_flds}
\alias{sync_flds}
\title{Synchornize fields in dataframe to match a database table}
\usage{
sync_flds(df, targ_db, tablename, drop_cols = TRUE)
}
\arguments{
\item{df}{\itemize{
\item a dataframe
}}

\item{targ_db}{\itemize{
\item path to our target accdb file
}}

\item{tablename}{\itemize{
\item the table name in the target databae to query against
}}

\item{drop_cols}{\itemize{
\item Boolean, Should extra names in the dataframe be dropped?
}}
}
\value{
dataframe
}
\description{
This function accepts a dataframe, a databae name and a table name
and adds and optionally removes fields from the dataframe to match
the names in the target table.  If drop_col=FALSE, extra fields in
}
\details{
the dataframe will remain, if drop_cols=TRUE, extra fields in the
dataframe will be removed.
}
\seealso{
get_trg_table_names
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
