% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfishr_utils.R
\name{truncate_slug}
\alias{truncate_slug}
\title{Truncate Slug}
\usage{
truncate_slug(slug, levels = 1)
}
\arguments{
\item{slug}{\itemize{
\item a string Often a human readable, unique idnetifier
}}

\item{levels}{\itemize{
\item a number, defaults to 1. It value indicated the
number of slug components (dash and string elements) to remove
from the input
}}
}
\value{
string
}
\description{
Truncate the last element from glis slug
}
\details{
Slug are used thorughout glis to uniquely identify individual
records.  Slugs are aways build a contenated string of key fields,
converted to lowercase and separated by dashes.  This function can
be used to create the slug of an objects parent by lopping off the
termal entity and dash.  It can be called recursively to obtian
slugs of grandparent or great-grandparent records if necessary.
}
\author{
R. Adam Cottrill
}
