% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfishr_utils.R
\name{uncount_tally}
\alias{uncount_tally}
\title{Uncount Tallies}
\usage{
uncount_tally(df, tally_col)
}
\arguments{
\item{df}{FN124 table fetched from GLIS using get_FN124()}

\item{tally_col}{Column name where the tally is stored; enter as a string}
}
\value{
dataframe
}
\description{
This function "uncounts" a tally, duplicating each row a number of times
equal to the value in the selected tally column. It is a simplified version
of the tidyr function uncount().
}
\examples{
LOA_IA21_TW1 <- get_FN124(list(prj_cd = "LOA_IA21_TW1"))
LOA_IA21_TW1_long <- uncount_tally(LOA_IA21_TW1, "SIZCNT")
}
