% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_template.R
\name{validate_glis_data}
\alias{validate_glis_data}
\title{Check data returned from api before inserting into template}
\usage{
validate_glis_data(glis_data)
}
\arguments{
\item{glis_data}{named list compiled from the glis api endpoints}
}
\value{
named list
}
\description{
This function runs a number of checks againts the glis data list
returned from teh api before it is inserted into the target
database.  If an error is found, the function is stopped and an
error message is presented to the user.
}
\author{
Arthur Bonsall \email{arthur.bonsall@ontario.ca}
}
