% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_filters.R
\name{add_special_filters}
\alias{add_special_filters}
\title{Add special fitler options to seelected endpoints}
\usage{
add_special_filters(endpoint_name, values)
}
\arguments{
\item{endpoint_name}{\itemize{
\item the name of api endpoint
}}

\item{values}{\itemize{
\item dataframe with existing filter attributes
}}
}
\value{
list
}
\description{
Some of the the api endpoints accept filter arguments that are not
captured in the automatically generated swagger docus.  These need
to be added so they can be presented to glfishr users.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
