% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Angler_Answers.R
\name{get_Angler_Answers}
\alias{get_Angler_Answers}
\title{Get Angler Answers from Creel Portal API}
\usage{
get_Angler_Answers(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for Angler Answers. Angler
Answer records contain the multiple choice answers to any extra
questions that were asked during creels. Each answer has two parts: the
Answer_Number and Answer_Text. The Answer_Number is recorded in the FN121
table in the ANGOP# fields, where the digit in the ANGOP# field corresponds
with the Question_Number. The Answer_Text is the text that corresponds
with the Answer_Number. This function takes an optional filter list which
can be used to return records based on the answer number, text, and
several attributes of the project such as project code,
or part of the project code, lake, first year, last year, etc.
}
\details{
Use \code{show_filters("angler_answers")} to see the full list of available filter
keys (query parameters). Refer to \url{https://intra.glis.mnr.gov.on.ca/creels/api/v1/swagger/}
and filter by "angler_answers" for additional information.
}
\examples{

answer_05_15 <- get_Angler_Answers(list(year__gte = 2005, year__lte = 2015))

boat_answers <- get_Angler_Answers(list(answer_text__like = "boat"))

manitoulin_answers <- get_Angler_Answers(list(lake = "HU", prj_cd__like = "120"))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
