% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Angler_Questions.R
\name{get_Angler_Questions}
\alias{get_Angler_Questions}
\title{Get Angler Questions from Creel Portal API}
\usage{
get_Angler_Questions(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for Angler Questions. Angler
Question records contain the extra questions that were asked during
creels. The question number corresponds with the ANGOP# fields in the
FN121 table. This function takes an optional filter list which can be
used to return records based on the question text, number, and several
attributes of the project such as project code, or part of the project
code, lake, first year, last year, etc.
}
\details{
Use ~show_filters("angler_questions")~ to see the full list of available filter
keys (query parameters)
}
\examples{

angler_questions <- get_Angler_Questions(list(lake = "ON", year__gte = 2012, year__lte = 2018))

angler_questions <- get_Angler_Questions(list(lake = "ER"))

angler_questions <- get_Angler_Questions(list(lake = "HU", prj_cd__like = "_001"))

angler_questions <- get_Angler_Questions(list(lake = "HU", protocol = "USA"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
