% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domain.R
\name{get_domain}
\alias{get_domain}
\title{Internal func to get the domain great lake portal api's}
\usage{
get_domain()
}
\value{
string
}
\description{
This funciton is intended to be used by all of the functions that #' call an
api endpoint exposed by the great lakes api endpoints.
Each of the apis. are likely to have slightly differnt urls (tfat/api/v1,
fn_portal/api/v1, ect), but a common domain. If the api domain changes in
the future, (ie - the domain name changes or the #' version changes) we just
need to update the value retruned by this function.
}
\details{
To use a local version of glfishr api during developement, set the
local environment variable to True:\preformatted{>  Sys.setenv(FN_PORTAL_DEV=TRUE)
}

By default, it is assumed that the api is being served on #' localhost and\preformatted{ >  Sys.setenv(FN_PORTAL_HOST='<some-ip-address>')
 > Sys.setenv(FN_PORTAL_PORT=8080)
}
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
