% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN022.R
\name{get_FN022}
\alias{get_FN022}
\title{Get FN022 - Season Strata Data FN_Portal API}
\usage{
get_FN022(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE,
  add_year_col = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}

\item{add_year_col}{\itemize{
\item should a 'year' column be added to the
returned dataframe?  This argument is ignored if the data frame
does not contain a 'prj_cd' column.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN022
records. FN022 records contain information about the defined season
associated with a project.
}
\details{
Use \code{show_filters("fn022")} to see the full list of available filter
keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/}
and filter by "fn022" for additional information.
}
\examples{

fn022 <- get_FN022(list(lake = "ON", year = 2012))

fn022 <- get_FN022(list(prj_cd = "LHA_IA19_812"))
fn022 <- get_FN022(list(prj_cd = "LHA_IA19_812"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
