% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN026_Subspace.R
\name{get_FN026_Subspace}
\alias{get_FN026_Subspace}
\title{Get FN026_subspace - Spatial Strata Data FN_Portal API}
\usage{
get_FN026_Subspace(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN026_Subspace
records. FN026_Subspace records contain information about the lower level
spatial strata associated with a project.
}
\details{
Use \code{show_filters("fn026subspace")} to see the full list of available filter
keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/}
and filter by "fn026subspace" for additional information.
}
\examples{

fn026_subspace <- get_FN026_Subspace(list(lake = "ON", year = 2012))

fn026_subspace <- get_FN026_Subspace(list(prj_cd = "LHA_IA19_812"))
fn026_subspace <- get_FN026_Subspace(list(prj_cd = "LHA_IA19_812"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
