% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121.R
\name{get_FN121}
\alias{get_FN121}
\title{Get FN121 - Net set data from FN_Portal API}
\usage{
get_FN121(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN121
records. FN121 records contain information about net sets or more
generally sampling events in OMNR netting projects.  The FN121
records contain information like set and lift date and time, effort
duration, gear, site depth and location.  This function takes an
optional filter list which can be used to return record based on
several attributes of the net set including set and lift date and
time, effort duration, gear, site depth and location as well as
attributes of the projects they are associated with such project
code, or part of the project code, lake, first year, last year,
protocol, etc.
}
\details{
Use ~show_filters("fn121")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/
and filter by "fn121" for additional information.
}
\examples{

fn121 <- get_FN121(list(lake = "ON", year = 2012))
fn121 <- get_FN121(list(
  lake = "ER", protocol = "TWL",
  sidep0__lte = 20, year__gte = 2010
))
filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN")
)
fn121 <- get_FN121(filters)
fn121 <- get_FN121(list(lake = "HU", prj_cd__like = "_003"))
fn121 <- get_FN121(list(lake = "HU", prj_cd__like = "_003"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
