% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121_Limno.R
\name{get_FN121_Limno}
\alias{get_FN121_Limno}
\title{Get FN121_Limno - Net set limnological data from FN_Portal API}
\usage{
get_FN121_Limno(
  filter_list = list(),
  with_121 = FALSE,
  show_id = FALSE,
  to_upper = TRUE
)
}
\arguments{
\item{filter_list}{list}

\item{with_121}{When 'FALSE', the default, only the limnology fields
from the FN121 table are returned. To return the whole FN121 table
(excluding trapnet, weather, and trawl fields), use 'with_121 = TRUE'.}

\item{show_id}{When 'FALSE', the default, the 'slug'
field is hidden from the data frame. To return this field
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?  Defaults to TRUE to match the FISHNET-II  convention
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN121limno
records. FN121limno records contain limnological data associated
with individual net sets. There is a 1:1 relationship between a
net set and a FN121Lino record. The FN121limno record contains
information like dissolved oxygen at the surface or at the gear at
the start and/or end of a sampling event. This function takes an
optional filter list which can be used to return records based on
several attributes of the limnological data as well as attributes
of the net set including set and lift date and
time, effort duration, gear, site depth and location as well as
attributes of the projects they are associated with such project
code, or part of the project code, lake, first year, last year,
protocol, etc.  The limnological data can be joined back to the
associated net set in R using the function \code{\link{join_fn_fields}}
}
\details{
Use ~show_filters("fn121limno")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/
and filter by "fn121limno" for additional information.
}
\examples{

# see the available filters that contain the string 'o2' (oxygen)
show_filters("fn121limno", "o2")

fn121_limno <- get_FN121_Limno(list(lake = "ER", year = 2019, o2surf0__not_null = TRUE))
fn121_limno <- get_FN121_Limno(list(lake = "ER", year = 2019, o2surf0__not_null = TRUE), with_121 = TRUE)
fn121_limno <- get_FN121_Limno(list(lake = "ER", year = 2019, o2surf0__not_null = TRUE, mu_type = "moe"), with_121 = TRUE)
fn121_limno <- get_FN121_Limno(list(lake = "ER", year = 2019, o2surf0__not_null = TRUE), show_id = TRUE)

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
