% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121_Trapnet.R
\name{get_FN121_Trapnet}
\alias{get_FN121_Trapnet}
\title{Get FN121_Trapnet - Trapnet data from FN_Portal API}
\usage{
get_FN121_Trapnet(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for fn121trapnet
records. fn121trapnet records contain details typically collected as part
of the Ontario NSCIN protocol: bottom type, cover type, vegetation,
and the angle, use, and distance offshore of the trap net leader.
Other relevant details for each SAM are found in the FN121 table.
This function takes an optional filter list which can be used to return records
based on attributes of the SAM including site depth, start and end date
and time, effort duration, gear, and location as well as attributes of the projects
they are associated with such as project code, or part of the project code, lake,
first year, last year, protocol, etc. This function can also take filters related to
the bottom, cover, and vegetation types, and the angle/length of the leader.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn121trapnet_list
for the full list of available filter keys (query parameters)
}
\examples{

TODO: Update with relevant examples when more data exists in the portal

fn121_trapnet <- get_FN121_Trapnet(list(lake = "ON", year = 2022))
fn121_trapnet <- get_FN121_Trapnet(list(lake = "ON", year = 2022), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
