% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121_Trapnet.R
\name{get_FN121_Trapnet}
\alias{get_FN121_Trapnet}
\title{Get FN121_Trapnet - Trapnet data from FN_Portal API}
\usage{
get_FN121_Trapnet(
  filter_list = list(),
  with_121 = FALSE,
  show_id = FALSE,
  to_upper = TRUE
)
}
\arguments{
\item{filter_list}{list}

\item{with_121}{When 'FALSE', the default, only the trapnet fields
from the FN121 table are returned. To return the whole FN121 table
(excluding limnology, weather, and trawl fields), use 'with_121 = TRUE'.}

\item{show_id}{When 'FALSE', the default, the 'slug'
field is hidden from the data frame. To return this field
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for fn121trapnet
records. fn121trapnet records contain details typically collected as part
of the Ontario NSCIN protocol: bottom type, cover type, vegetation,
and the angle, use, and distance offshore of the trap net leader.
Other relevant details for each SAM are found in the FN121 table.
This function takes an optional filter list which can be used to return records
based on attributes of the SAM including site depth, start and end date
and time, effort duration, gear, and location as well as attributes of the projects
they are associated with such as project code, or part of the project code, lake,
first year, last year, protocol, etc. This function can also take filters related to
the bottom, cover, and vegetation types, and the angle/length of the leader.
}
\details{
Use ~show_filters("fn121trapnet")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/
and filter by "fn121trapnet" for additional information.
}
\examples{

fn121_trapnet <- get_FN121_Trapnet(list(protocol="NSCIN",
bottom_type="GP", year=2022))

fn121_trapnet <- get_FN121_Trapnet(list(protocol="NSCIN",
bottom_type="GP", mu_type="qma"), with_121=TRUE)

fn121_trapnet <- get_FN121_Trapnet(list(protocol="NSCIN",
bottom_type="GP"), show_id=TRUE)

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
