% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN121_Weather.R
\name{get_FN121_Weather}
\alias{get_FN121_Weather}
\title{Get FN121_Weather - Weather data from FN_Portal API}
\usage{
get_FN121_Weather(
  filter_list = list(),
  with_121 = FALSE,
  show_id = FALSE,
  to_upper = TRUE
)
}
\arguments{
\item{filter_list}{list}

\item{with_121}{When 'FALSE', the default, only the weather fields
from the FN121 table are returned. To return the whole FN121 table
(excluding limnology, trapnet, and trawl fields), use 'with_121 = TRUE'.}

\item{show_id}{When 'FALSE', the default, the 'slug'
field is hidden from the data frame. To return this field
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for fn121weather
records. fn121weather records contain air temperature, precipitation,
wind speed and direction, cloud cover, and wave height data collected
at time/location 0 and/or 1. Other relevant details for each SAM are
found in the FN121 table. This function takes an optional filter list which can
be used to return records based on attributes of the SAM including site
depth, start and end date and time, effort duration, gear, and location
as well as attributes of the projects they are associated with such as project
code, or part of the project code, lake, first year, last year,
protocol, etc. This function can also take filters related to weather.
}
\details{
Use ~show_filters("fn121weather")~ to see the full list of available filter
keys (query parameters). Refer to https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/
and filter by "fn121weather" for additional information.
}
\examples{

fn121_weather <- get_FN121_Weather(list(lake = "ER", year = 2018))
fn121_weather <- get_FN121_Weather(list(lake = "ER", year = 2018), with_121 = TRUE)
fn121_weather <- get_FN121_Weather(list(lake = "ER", year = 2018, mu_type = "ypmu"), with_121 = TRUE)
fn121_weather <- get_FN121_Weather(list(lake = "ER", year = 2018), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
