% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN125Tags.R
\name{get_FN125Tags}
\alias{get_FN125Tags}
\title{Get FN125Tags - Tagging data from FN_Portal API}
\usage{
get_FN125Tags(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN125Tags
records. FN125Tags records contain information about the
individual tags applied to or recovered from on a sampled fish.
Historically, tag data was stored in three related fields -
TAGDOC, TAGSTAT and TAGID.  This convention is fine as long a
single biological sample only has a one tag. In recent years, it
has been come increasingly common for fish to have multiple tags,
or tag types associated with indiviudal sampling events. FN125Tag
accomodates those events.  This function takes an optional filter
list which can be used to return records based on several
different attributes of the tag (tag type, colour, placement,
agency, tag stat, and tag number) as well as, attributes of the
sampled fish such as the species, or group code, or attributes of
the effort, the sample, or the project(s) that the samples were
collected in.
}
\details{
See
http://http://10.167.37.157/fn_portal/redoc/#operation/fn_125tags_list
for the full list of available filter keys (query parameters)
}
\examples{

fn125Tags <- get_FN125Tags(list(lake='ON', year=2012, spc='334',gear='GL'))

filters <- list(lake='ER',
                protocol='TWL',
                spc_in=c('331', '334'),
                sidep_lte=20)
fn125Tags <- get_FN125Tags(filters)

filters <- list(lake='SU',
           prj_cd__in=c('LSA_IA15_CIN','LSA_IA17_CIN'),
           eff='051',
           spc='091')
fn125Tags <- get_FN125Tags(filters)

filters <- list(lake='HU', spc='076', grp='55')
fn125Tags <- get_FN125Tags(filters)

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
