% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN126.R
\name{get_FN126}
\alias{get_FN126}
\title{Get FN126 - Diet data from FN_Portal API}
\usage{
get_FN126(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN126
records. FN126 records contain the counts of identifiable items in
found in the stomachs of sampled fish.  This function takes an
optional filter list which can be used to return records based on
several different attributes of the diet item (taxon, taxon_like),
as well as, attributes of the sampled fish such as the species, or
group code, or attributes of the effort, the sample, or the
project(s) that the samples were collected in.
}
\details{
Use \code{show_filters("fn126")} to see the full list of available
filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/}
and filter by "fn126" for additional information.

The TAXON field in the FN126 table returns HHFAU (Haliburton
Hastings Fishery Assessment Unit) codes by default if they
exist. To return only ITIS (Integrated Taxonomic Information
System) values, specify 'itiscode=TRUE' in the list of filters
passed into the get_FN126 function.
}
\examples{

fn126 <- get_FN126(list(lake = "ON", year = 2012, spc = "334", gr = "GL"))

filters <- list(
  lake = "ER",
  protocol = "TWL",
  spc = c("331", "334"),
  sidep0__lte = 20
)
fn126 <- get_FN126(filters)

filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"),
  eff = "051",
  spc = "091"
)
fn126 <- get_FN126(filters)

filters <- list(lake = "HU", spc = "076", grp = "55")
fn126 <- get_FN126(filters)
fn126 <- get_FN126(filters, show_id = TRUE)

# TAXON will contain ITIS values rather than HHFAU codes:
filters <- list(lake = "HU", spc = "076", grp = "55", itiscode = TRUE)
fn126 <- get_FN126(filters)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
