% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN126.R
\name{get_FN126}
\alias{get_FN126}
\title{Get FN126 - Diet data from FN_Portal API}
\usage{
get_FN126(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN126
records. FN126 records contain the counts of identifiable items in
found in the stomachs of sampled fish.  This function takes an
optional filter list which can be used to return records based on
several different attributes of the diet item (taxon, taxon_like),
as well as, attributes of the sampled fish such as the species, or
group code, or attributes of the effort, the sample, or the
project(s) that the samples were collected in.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn_126_list
for the full list of available filter keys (query parameters)
}
\examples{

fn126 <- get_FN126(list(lake = "ON", year = 2012, spc = "334", gear = "GL"))

filters <- list(
  lake = "ER",
  protocol = "TWL",
  spc = c("331", "334"),
  sidep__lte = 20
)
fn126 <- get_FN126(filters)

filters <- list(
  lake = "SU",
  prj_cd = c("LSA_IA15_CIN", "LSA_IA17_CIN"),
  eff = "051",
  spc = "091"
)
fn126 <- get_FN126(filters)

filters <- list(lake = "HU", spc = "076", grp = "55")
fn126 <- get_FN126(filters)
fn126 <- get_FN126(filters, show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
