% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FN127.R
\name{get_FN127}
\alias{get_FN127}
\title{Get FN127 - Diet data from FN_Portal API}
\usage{
get_FN127(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for FN127 - age
estimate/interpretations.  This function takes an optional filter
list which can be used to return records based on several
different attributes of the age estimate such as the assigned age,
the aging structure, confidence, number of complete annuli and
edge code, or whether or not it was identified as the 'preferred'
age for this fish. Additionally, filters can be applied to
select age estimates based on attributes of the sampled fish such
as the species, or group code, or attributes of the effort, the
sample, or the project(s) that the samples were collected in.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn_127_list
for the full list of available filter keys (query parameters)
}
\examples{

fn127 <- get_FN127(list(lake='ON', year=2012, spc='334', gear='GL'))

filters <- list(lake='ER',
                protocol='TWL',
                spc=c('331', '334'),
                year=2010,
                sidep_lte=20)
fn127 <- get_FN127(filters)

filters <- list(lake='SU',
           prj_cd_in=c('LSA_IA15_CIN','LSA_IA17_CIN'),
           eff='051',
           spc='091')
fn127 <- get_FN127(filters)

filters <- list(lake='HU', spc='076', grp='55')
fn127 <- get_FN127(filters)

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
