% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gears.R
\name{get_gear}
\alias{get_gear}
\title{Get Gear List from the FN_Portal API}
\usage{
get_gear(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function returns basic details about fishing gear(s) from the
FN_Prortal api. Use get_gear_process_types() for more detail about
the corresponding process type
}
\details{
Use \code{show_filters("gear")} to see the full list of available
filter keys (query parameters). Refer to
\url{https://intra.glis.mnr.gov.on.ca/fn_portal/api/v1/swagger/}
and filter by "gear" for additional information.
}
\examples{

GL10 <- get_gear(list(gr = "GL10"))
GL_n <- get_gear(list(gr = c("GL10", "GL21")))
GL1x <- get_gear(list(gr__like = "GL1"))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
