% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu_list.R
\name{get_mu_list}
\alias{get_mu_list}
\title{Get list of lake management unit types from GLIS API}
\usage{
get_mu_list(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{should the names of the dataframe be converted to
upper case?}
}
\value{
dataframe
}
\description{
Get list of lake management unit types from GLIS API
}
\examples{
get_mu_list()
get_mu_list(list(lake = "HU"))
get_mu_list(list(lake = "HU", mu_type = "basin"))
}
