% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC024.R
\name{get_SC024}
\alias{get_SC024}
\title{Get SC024 - Period Type Data Creel_Portal API}
\usage{
get_SC024(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for SC024 records. SC024
records contain information about the period used in a creel
survey. Often the fishing day is split into morning and afternoon,
and in some instances the periods change between seasons.
}
\details{
Use ~show_filters("sc024")~ to see the full list of available filter
keys (query parameters)
}
\examples{

sc024 <- get_SC024(list(lake = "ON", year = 2012))

sc024 <- get_SC024(list(prj_cd = "LOA_SC12_002"))
sc024 <- get_SC024(list(prj_cd = "LOA_SC12_002"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
