% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC026_Subspace.R
\name{get_SC026_Subspace}
\alias{get_SC026_Subspace}
\title{Get SC026_Subspace - Spatial Strata Data Creel_Portal API}
\usage{
get_SC026_Subspace(filter_list = list(), show_id = FALSE, to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{When 'FALSE', the default, the 'id' and 'slug'
fields are hidden from the data frame. To return these columns
as part of the data frame, use 'show_id = TRUE'.}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for SC026_Subspace
records. SC026_Subspace records contain information about the lower level
spatial strata associated with a project.
}
\details{
See
http://10.167.37.157/creels/api/v1/redoc/#operation/sc026subspace_list
for the full list of available filter keys (query parameters)
}
\examples{

sc026_subspace <- get_SC026_Subspace(list(lake = "ON", year = 2012))

sc026_subspace <- get_SC026_Subspace(list(prj_cd = "LHA_SC03_220"))
sc026_subspace <- get_SC026_Subspace(list(prj_cd = "LHA_SC03_220"), show_id = TRUE)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
