% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get Species list data from Common API}
\usage{
get_species(filter_list = list(), to_upper = TRUE)
}
\arguments{
\item{filter_list}{list}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint to for Species list
available in Common. Only the fields spc, spc_nmco, spc_nmsc
and species_at_risk are returned unless detail=TRUE is specified
in the filter_list. The optional filter list can be used to return
species based on part of their common or scientific name.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/species_list_list
for the full list of available filter keys (query parameters)
}
\examples{

species <- get_species()
trout <- get_species(list(spc_nmco__like = "trout"))
goby <- get_species(list(spc = 366, detail = TRUE))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
