% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_positions.R
\name{get_tag_positions}
\alias{get_tag_positions}
\title{Get tag position options - A list of tag position codes used in GLIS}
\usage{
get_tag_positions(
  filter_list = list(),
  show_id = FALSE,
  to_upper = TRUE,
  record_count = FALSE
)
}
\arguments{
\item{filter_list}{list}

\item{show_id}{include the fields the 'id' and 'slug' in the
returned data frame}

\item{to_upper}{\itemize{
\item should the names of the dataframe be converted to
upper case?
}}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the api endpoint for tag position choices and returns
their labels, descriptions and whether they're in use. It fetches
the entire table of accepted tag positions - no other filter
parameters are currently available for this endpoint. Tag position is the
second character of TAGDOC.
}
\details{
See
https://intra.glis.mnr.gov.on.ca/common/tag_positions/
for the full list of tag position options
}
\examples{

tag_positions <- get_tag_positions()
tag_position_slugs <- get_tag_positions(show_id = TRUE)
}
\author{
Rachel Henderson \email{rachel.henderson@ontario.ca}
}
