% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfat_encounters.R
\name{get_tfat_encounters}
\alias{get_tfat_encounters}
\title{Get TFAT Tag Encounters}
\usage{
get_tfat_encounters(filter_list = list(), record_count = FALSE)
}
\arguments{
\item{filter_list}{list}

\item{record_count}{\itemize{
\item should data be returned, or just the number
of records that would be returned given the current filters.
}}
}
\value{
dataframe
}
\description{
This function accesses the TFAT api endpoint for tag encounters - sampling
events were tags were either applied or re-captured during OMNR field
projects. (excluding CWTs). The TFAT encounter information contains all of
the information required to analyze  tagging data, but they are not the
definitive record for biological samples, and do not contain any information
about other fish caught in the same sampling event.  See Project Tracker for
more information about tagging projects and FN_Portal(and other master
databases) to access the complete dataset.
}
\details{
This function takes an optional filter list which can be used to return
records based on several attributes of the tagged fish such as species, size,
sex, or maturity, the tag attributes such as colour, placement, tag type, or
tag status (either tagged or recovered), or attributes of the projects in
which the encounters occurred such as lake or year the projects were run.
}
\examples{

# fish that had tags on capture in Lake Huron between 2012 and 2014
encounters <- get_tfat_encounters(list(
  lake = "HU", year__gte = 2012, year__lte =
    2014, tagstat = "C"
))

# Fish that were sampled in Lake Superior and had yellow OMNR tag applied,
# or had a yellow omnr tag on applied
filters <- list(lake = "SU", tagdoc__endswith = "012", tag_stat = "A")
encounters <- get_tfat_encounters(filters)

# recovered walleye or musky
filters <- list(spc = c("334", "132"), tagstat = "C", year__gte = 2015, year__lte = 2016)
encounters <- get_tfat_encounters(filters)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
