% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_filters.R
\name{show_filters}
\alias{show_filters}
\title{Show filters available on the provided api endpoint}
\usage{
show_filters(endpoint = "", filter_like = "")
}
\arguments{
\item{endpoint}{\itemize{
\item the name of the api endpoint
}}

\item{filter_like}{\itemize{
\item a partial string to match filter against
}}
}
\description{
Most of the fn_portal API endpoints have a large number of
available filters.  This function takes and api endpoint name and
prints out all of the available filters.  If 'filter_like' is
provided, only those filters that contain the provided string are
printed. If no endpoint name is provided, a list of available
endpoints will be displayed.
}
\details{
See
http://10.167.37.157/fn_portal/redoc/#operation/fn_011_list for the
full list of available filter keys (query parameters)
}
\examples{
show_filters("fn125")
show_filters("fn125", filter_like = "prj")
show_filters("foo")
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
